% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenologyDistributions.R
\name{dO}
\alias{dO}
\alias{pO}
\alias{qO}
\alias{rO}
\alias{dOk1}
\alias{qOk1}
\alias{pOk1}
\alias{rOk1}
\alias{dC}
\alias{pC}
\alias{rC}
\alias{qC}
\alias{Pt}
\alias{dT}
\alias{pT}
\alias{qT}
\alias{rT}
\alias{rCkN}
\alias{dCkN}
\alias{pCkN}
\alias{qCkN}
\alias{dR}
\alias{rR}
\alias{dPNt}
\alias{rPNt}
\title{Distribution functions}
\usage{
dO(x, mu_O, sigma_O, minResponse = 0, maxResponse = 365, type = c("GP", "BB"))

pO(q, mu_O, sigma_O, minResponse = 0, maxResponse = 365, type = c("GP", "BB"))

qO(p, mu_O, sigma_O, minResponse = 0, maxResponse = 365, type = c("GP", "BB"))

rO(n, mu_O, sigma_O, minResponse = 0, maxResponse = 365, type = c("GP", "BB"))

dOk1(
  x,
  N,
  mu_O,
  sigma_O,
  minResponse = 0,
  maxResponse = 365,
  type = c("GP", "BB")
)

qOk1(
  p,
  N,
  mu_O,
  sigma_O,
  minResponse = 0,
  maxResponse = 365,
  type = c("GP", "BB")
)

pOk1(
  q,
  N,
  mu_O,
  sigma_O,
  minResponse = 0,
  maxResponse = 365,
  type = c("GP", "BB")
)

rOk1(
  n,
  N,
  mu_O,
  sigma_O,
  minResponse = 0,
  maxResponse = 365,
  type = c("GP", "BB")
)

dC(
  x,
  mu_O,
  sigma_O,
  mu_D,
  sigma_D,
  minResponse = 0,
  maxResponse = 365,
  type = c("GP", "BB")
)

pC(
  q,
  mu_O,
  sigma_O,
  mu_D,
  sigma_D,
  minResponse = 0,
  maxResponse = 365,
  type = c("GP", "BB")
)

rC(
  n,
  mu_O,
  sigma_O,
  mu_D,
  sigma_D,
  minResponse = 0,
  maxResponse = 365,
  type = c("GP", "BB")
)

qC(
  p,
  mu_O,
  sigma_O,
  mu_D,
  sigma_D,
  minResponse = 0,
  maxResponse = 365,
  type = c("GP", "BB")
)

Pt(
  x,
  mu_O,
  sigma_O,
  mu_D,
  sigma_D,
  minResponse = 0,
  maxResponse = 365,
  type = c("GP", "BB")
)

dT(
  x,
  mu_O,
  sigma_O,
  mu_D,
  sigma_D,
  minResponse = 0,
  maxResponse = 365,
  type = c("GP", "BB")
)

pT(
  q,
  mu_O,
  sigma_O,
  mu_D,
  sigma_D,
  minResponse = 0,
  maxResponse = 365,
  type = c("GP", "BB")
)

qT(
  p,
  mu_O,
  sigma_O,
  mu_D,
  sigma_D,
  minResponse = 0,
  maxResponse = 365,
  type = c("GP", "BB")
)

rT(
  n,
  mu_O,
  sigma_O,
  mu_D,
  sigma_D,
  minResponse = 0,
  maxResponse = 365,
  type = c("GP", "BB")
)

rCkN(
  n,
  N,
  mu_O,
  sigma_O,
  mu_D,
  sigma_D,
  minResponse = 0,
  maxResponse = 365,
  type = c("GP", "BB")
)

dCkN(
  x,
  N,
  mu_O,
  sigma_O,
  mu_D,
  sigma_D,
  minResponse = 0,
  maxResponse = 365,
  type = c("GP", "BB")
)

pCkN(
  q,
  N,
  mu_O,
  sigma_O,
  mu_D,
  sigma_D,
  minResponse = 0,
  maxResponse = 365,
  type = c("GP", "BB")
)

qCkN(
  p,
  N,
  mu_O,
  sigma_O,
  mu_D,
  sigma_D,
  minResponse = 0,
  maxResponse = 365,
  type = c("GP", "BB")
)

dR(
  x,
  N,
  mu_O,
  sigma_O,
  mu_D,
  sigma_D,
  minResponse = 0,
  maxResponse = 365,
  type = c("GP", "BB")
)

rR(
  n,
  N,
  mu_O,
  sigma_O,
  mu_D,
  sigma_D,
  minResponse = 0,
  maxResponse = 365,
  type = c("GP", "BB")
)

dPNt(
  x,
  t,
  mu_O,
  sigma_O,
  mu_D,
  sigma_D,
  minResponse = 0,
  maxResponse = 365,
  type = c("GP", "BB")
)

rPNt(
  n,
  t,
  mu_O,
  sigma_O,
  mu_D,
  sigma_D,
  minResponse = 0,
  maxResponse = 365,
  type = c("GP", "BB")
)
}
\arguments{
\item{x}{data}

\item{mu_O}{mu 0}

\item{sigma_O}{sigma 0}

\item{minResponse}{min of the response}

\item{maxResponse}{max of the response}

\item{type}{the type}

\item{q}{the quantile}

\item{p}{the probability}

\item{n}{the number of samples}

\item{N}{the population size.}

\item{mu_D}{mu D}

\item{sigma_D}{sigma D}

\item{t}{the parameter t}
}
\value{
The results.
}
\description{
These are distribution functions with 'd' (density), 'r' (sampling), 'q' (quantiles), and 'p' (probability) options. Most functions share the same input parameters.
}
