% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenologyInference.R
\name{E.CkN}
\alias{E.CkN}
\title{Title}
\usage{
E.CkN(
  N,
  mu_O,
  sigma_O,
  mu_D,
  sigma_D = NA,
  minResponse = 0,
  maxResponse = 365,
  type = c("GP", "BB"),
  threshApprox = NA
)
}
\arguments{
\item{N}{}

\item{mu_O}{}

\item{sigma_O}{}

\item{mu_D}{}

\item{sigma_D}{}

\item{minResponse}{}

\item{maxResponse}{}

\item{type}{}

\item{threshApprox}{}
}
\description{
Title
}
