% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenologyInference.R
\name{E.O}
\alias{E.O}
\title{Title}
\usage{
E.O(mu_O, sigma_O, minResponse = 0, maxResponse = 365, type = c("GP", "BB"))
}
\arguments{
\item{mu_O}{}

\item{sigma_O}{}

\item{minResponse}{}

\item{maxResponse}{}

\item{type}{}
}
\description{
Title
}
