% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenologyInference.R
\name{E.T}
\alias{E.T}
\title{Title}
\usage{
E.T(mu_O, mu_D, minResponse = 0, maxResponse = 365, type = c("GP", "BB"))
}
\arguments{
\item{mu_O}{}

\item{mu_D}{}

\item{minResponse}{}

\item{maxResponse}{}

\item{type}{}
}
\description{
Title
}
