% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenologyInference.R
\name{phenologyExpectationFunctions}
\alias{phenologyExpectationFunctions}
\alias{E.C}
\alias{E.CkN}
\alias{E.D}
\alias{E.O}
\alias{E.Ok1}
\alias{E.T}
\title{Functions that calculate the expected value of the associated phenological random variable:}
\usage{
E.C(mu_O, mu_D = NA, minResponse = 0, maxResponse = 365, type = c("GP", "BB"))

E.CkN(
  N,
  mu_O,
  sigma_O,
  mu_D,
  sigma_D = NA,
  minResponse = 0,
  maxResponse = 365,
  type = c("GP", "BB"),
  threshApprox = NA
)

E.D(
  mu_O,
  sigma_O,
  mu_D,
  sigma_D = NA,
  minResponse = 0,
  maxResponse = 365,
  type = c("GP", "BB")
)

E.O(mu_O, sigma_O, minResponse = 0, maxResponse = 365, type = c("GP", "BB"))

E.Ok1(
  N,
  mu_O,
  sigma_O,
  minResponse = 0,
  maxResponse = 365,
  type = c("GP", "BB"),
  threshApprox = NA
)

E.T(mu_O, mu_D, minResponse = 0, maxResponse = 365, type = c("GP", "BB"))
}
\arguments{
\item{mu_O}{Mean onset time}

\item{mu_D}{Mean duration time}

\item{minResponse}{Minimum value of the response (e.g., day of year); must be set to 0 under current implementation (default = 0)}

\item{maxResponse}{Maximum value of the response (e.g., day of year); typically 365 for Gregorian calendar (default = 365)}

\item{type}{The model type, either BB (beta onset, beta duration) or GP (Gaussian process with a shared standard deviation for onset and cessation and a constant duration) (default = "GP")}

\item{N}{The population size for estimation of extreme events}

\item{sigma_O}{Standard deviation for the onset time distribution}

\item{sigma_D}{Standard deviation for the duration time distribution}
}
\value{
The expected value of the associated random variable and input parameters
}
\description{
O: phenological onset times
}
\details{
D: phenophase durations

C: phenological cessation times

Ok1: first onset (order k = 1)

CkN: last cessation (order k = N, the population size)

T: observed specimen collection times, e.g., day of year for a yearly time period

Most functions share a subset of the same input parameters, but if you use a function that does not require all the parameters, provide only the needed parameters.

The function name is created by prepending "E." in front of one of the above variables names.
}
\examples{

#Set the mean onset time (day of year)
mu_O = 100
#Set the standard deviation of the distribution of onset times
sigma_O = 7
#set the mean duration of the phenophase (days)
mu_D = 30
#Set the standard deviation of the distribution of durations
sigma_D = 7
#Set the population size
N = 100000

#Calculate the expected cessation time for the beta onset, beta duration (BB) model
eC = E.C(mu_O=mu_O, mu_D=mu_D, type="BB")

#Calculate the expected last cessation time under the BB model
eCkN = E.CkN(N=N, mu_O=mu_O, sigma_O=sigma_O, mu_D=mu_D, sigma_D=sigma_D, type="BB") 

}
