% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{getHyperparametersViaQuantiles}
\alias{getHyperparametersViaQuantiles}
\title{Title}
\usage{
getHyperparametersViaQuantiles(
  responseDataForPrior,
  scale = 1,
  lowerQuantile = 0.1,
  upperQuantile = 0.9
)
}
\arguments{
\item{responseDataForPrior}{}

\item{scale}{}

\item{lowerQuantile}{}

\item{upperQuantile}{}
}
\description{
Title
}
