% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenologyVisualization.R
\name{makeMultistagePosteriorPredictivePlot}
\alias{makeMultistagePosteriorPredictivePlot}
\title{Make a posterior predictive plot from a "multistage-full" dataset analysis focused on one covariate}
\usage{
makeMultistagePosteriorPredictivePlot(
  stanResult,
  responseData,
  responseVariableName = "DOY",
  targetCovariateName,
  covariateData,
  stageData,
  nReps = 100,
  nXs = 101,
  nStages,
  slice = 0.25,
  minResponse = 0,
  maxResponse = 365,
  y_pred = FALSE,
  nChains = 4,
  nIts = 1000,
  smooth = c("GAM", "LOESS", "SPLINE", "NONE")
)
}
\arguments{
\item{stanResult}{The output from the function 'runStanPhenology' with a 'multistge-full' model}

\item{responseData}{A vector of the response data}

\item{responseVariableName}{The name of the response variable (default: DOY)}

\item{targetCovariateName}{A string with the name of the target covariate to be graphed on the x-axis of the output plot}

\item{covariateData}{A data frame with the onset covariate data. Columns are labeled with covariate names and rows have the covariate data. Rows should correspond to rows in the responseData vector.}

\item{stageData}{A vector of the stage for each item in the responseData vector}

\item{nReps}{The number of replicates to be used during the Monte Carlo simulation for each value of the target covariate. (default: 100)}

\item{nXs}{The number of intervals to be used for the target covariate. (default: 100)}

\item{nStages}{The number of stages.}

\item{slice}{he relative position where time slices will be made (also at 1 - slice). Plots of phenological distributions are made at the time slices. (default: 0.25)}

\item{minResponse}{Must be 0. (default: 0)}

\item{maxResponse}{The maxiumum possible value of the observed collection times. (default: 365)}
}
\value{
a ggplot2 graphics object
}
\description{
Make a graphic of the posterior predictive values of the onsets of multiple stages including 95\% credible intervals marginalized for a single covariate displayed along the x-axis.
}
\examples{
\donttest{
##Run a full multistage analysis with SIMULATION PARAMETERS themselves simulated and SIMULATED COVARIATES with MULTIPLE STAGES followed by a posterior predictive plot of the results
#
library(dplyr)
library(posterior)
 #Set the parameters
 nStages = 4
 nCovariates = 3
 nXs = 101
 nReps = 100
 n = 500
 minResponse = 0             #The minimum observed time should always be at least 0
 maxResponse = 365
 simulatedData = simulateMultistageData(n=n, nStages=nStages,nCovariates=nCovariates)

 #Plotting
     #Set which covariate is the x-axis in the plot
 targetCovariateIndex = 1

     #Set colors
 #stageColors = viridisLite::viridis(nStages+1)
 stageColors = RColorBrewer::brewer.pal(nStages+1, "Set1")

 trueModels = plotMultistageSimulation(simulatedData=simulatedData,
                  targetCovariateIndex=targetCovariateIndex,
                  stageColors=stageColors,
                  drawLatentOnset=FALSE,
                  drawObserved=TRUE,
                  drawTrueModel=FALSE,
                  drawInferredModel=FALSE,
                  shadeStage=TRUE,
                  minResponse=minResponse, maxResponse=maxResponse)
 dev.new()
 #Run inference with Stan
 stanResult =runStanPhenology(
     type="multistage-full",                             #model with many stages and covariates
     responseData=simulatedData$outputData$sampledTime,  #observed collection times
     stage=simulatedData$outputData$sampledStage,        #observed stage at collection time
     nStages=nStages,                                            #number of stages
     nOnsetCovariates=nCovariates,                       #number of onset covariates (same as duration)
     nDurationCovariates=nCovariates,                     #number of duration covariates (same as onset)
     onsetCovariateData=simulatedData$X,                 #covariate data (same as for duration)
     durationCovariateData=simulatedData$X,              #covariate data (same as for onset)
nReps=nReps,					    #MC replicates for posterior predictive distribution
nXs=nXs,					    #Number of intervals on x-axis of posterior predictive plot
     maxDiv=4000                                         #should be set to 0, but in case of the stray divergence, set high for this example
 )
###Make the posterior predictive plot
p = makeMultistagePosteriorPredictivePlot(stanResult=stanResult, 
		responseData=simulatedData$outputData$sampledTime, 
		targetCovariateName="cov1", 
		covariateData=simulatedData$X, 
		stageData=simulatedData$outputData$sampledStage, 
		nReps=nReps,	#this is the default for runStanPhenology
		nXs=nXs,	#this is the default for runStanPhenology
		smooth="GAM",
		nStages=nStages+1)
 
 for(i in 1:nStages) {
 	p = p + geom_abline(intercept = trueModels$trueIntercepts[i], slope = trueModels$trueSlopes[i], color = "black", linewidth = 0.5, linetype="dashed")
 }
 #Extract basic summary data
 probs = c(2.5, 97.5)
 measures=c("mean", "median", "sd", "mad")

 summary = print(
   stanResult$result$summary(
     variables = c(
       "sigma","anchor_d", "beta_d", "alpha_d", "anchor_o", "beta_o", "alpha_o"
      ),
     quantiles = ~ quantile2(., probs=probs/100),
     measures
   ),
   n = Inf
 ) \%>\% as.data.frame()

 #Set up vectors to store mean durations (means) and mean onsets (means_o) with low and high bounds of credible intervals.
 means = rep(0,nStages+1)
 mean_low = rep(0,nStages)
 mean_high = rep(0,nStages)

 #Extract slopes and intercepts (anchors) and overlay the inferred lines onto earlier plot
 for(j in 1:(nStages+1)) {
         beta = rep(0,nCovariates)
         beta_low = rep(0,nCovariates)
         beta_high = rep(0,nCovariates)
         varO = paste0("anchor_o[", j, "]")
         if(j<=nStages) {
                 varD = paste0("anchor_d[", j, "]")
                 means[j] = summary[summary$variable==varD, "mean"]
                 mean_low[j] = summary[summary$variable==varD, paste0("q",probs[1])]
                 mean_high[j] = summary[summary$variable==varD, paste0("q",probs[2])]
         }
         alpha = summary[summary$variable==varO, "mean"]
         alpha_low = summary[summary$variable==varO, paste0("q",probs[1])]
         alpha_high = summary[summary$variable==varO, paste0("q",probs[2])]

        for(i in 1:nCovariates) {
                 varO = paste0("beta_o[", j, ",", i, "]")
                 beta[i] = summary[summary$variable==varO, "mean"]
                 beta_low[i] = summary[summary$variable==varO, paste0("q",probs[1])]
                 beta_high[i] = summary[summary$variable==varO, paste0("q",probs[2])]
         }
        model = phenoCollectR:::true_marginal_line(alpha=alpha, beta=beta, mu=simulatedData$covariateMeans, Sigma=simulatedData$Sigma, j=targetCovariateIndex)
        model_low = phenoCollectR:::true_marginal_line(alpha=alpha_low, beta=beta_low, mu=simulatedData$covariateMeans, Sigma=simulatedData$Sigma, j=targetCovariateIndex)
        model_high = phenoCollectR:::true_marginal_line(alpha=alpha_high, beta=beta_high, mu=simulatedData$covariateMeans, Sigma=simulatedData$Sigma, j=targetCovariateIndex)
 	p = p + geom_abline(intercept = model$intercept, slope = model$slope, color = "red", linewidth = 0.5, linetype="dashed")
 }
print(p)
}
}
