% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenologyVisualization.R
\name{makePosteriorPredictivePlot}
\alias{makePosteriorPredictivePlot}
\title{Make a posterior predictive plot focussed on one covariate and marginalized across the other covariates}
\usage{
makePosteriorPredictivePlot(
  stanResult,
  responseData,
  responseVariableName = "DOY",
  targetCovariateName,
  onsetCovariateData,
  durationCovariateData,
  n_samples = 100,
  n_draws = 100,
  n_hist = 10000,
  n_bin = 100,
  resolution = 50,
  slice = 0.25,
  N = 1000,
  minResponse = 0,
  maxResponse = 365,
  adjustYLim = TRUE,
  smooth = c("GAM", "LOESS", "SPLINE", "NONE"),
  keepSigma = FALSE,
  makeHistograms = FALSE
)
}
\arguments{
\item{stanResult}{The output from the function 'runStanPhenology' with a 'full' model}

\item{responseData}{A vector of the response data}

\item{responseVariableName}{A vector of the names of the covariates}

\item{targetCovariateName}{A string with the name of the target covariate to be graphed on the x-axis of the output plot}

\item{onsetCovariateData}{A data frame with the onset covariate data. Can be obtained from the function 'preparePhenologyData'.}

\item{durationCovariateData}{A data frame with the duration covariate data. Can be obtained from the function 'preparePhenologyData'.}

\item{n_samples}{The number of covariate values to simulate for each value of the target covariate. Higher values give more accurate and more precise estimates. (default: 100)}

\item{n_draws}{The number of draws from the posterior distribution. Higher values give more accurate and more precise estimates. (default: 100)}

\item{n_hist}{The sample size of values to be used to approximate the onset, duration, and observed distributions. (default: 10000)}

\item{n_bin}{The number of bins to be used for the histograms. (default: 100)}

\item{resolution}{The number of equi-distant values of the target covariate to be used for graphing. Higher values give smoother curves, but take longer to render. (default: 50)}

\item{slice}{The relative position where time slices will be made (also at 1 - slice). Plots of phenological distributions are made at the time slices. (default: 0.25)}

\item{N}{The population size used to estimate phenological extremes. (default: 1000)}

\item{minResponse}{The minimum possible value of the response variable. (default: 0)}

\item{maxResponse}{The maximum possible value of the response variable. (default: 365)}

\item{adjustYLim}{Boolean flag specifying whether to set the limits of the y-axis adaptively. Use default. (default: TRUE)}

\item{smooth}{Define the type of smoothing, if any, to be used in the output plot. Possibilities are GAM, LOESS, SPLINE, and NONE. (default: GAM)}

\item{keepSigma}{Boolean flag indicating whether the sigma parameter should be included in the graphical output. (default: FALSE)}

\item{makeHistograms}{Boolean flag indicating whether histograms (vs. densities) at the time slices should be made. (default: FALSE)}
}
\value{
A ggplot2 graphics object
}
\description{
Make a graphic of the posterior predictive values of the extremes, onset, observed, and cessation values as well as sub-plots indication directions of change for onset, duration, and observed collection times.
}
\examples{
\donttest{
##get the file name with data for the blood root plant
file  =  getDatasetPath("Sanguinaria_canadensis")
## See documentation for more species:
help(getDatasetPath)
##define the covariate names - remove up to all but 1
vars = c("Latitude", "Year", "Elevation", "AnnualMonthlyAverageTemp", "SpringMonthlyAverageTemp"
         , "FirstQuarterMonthlyAverageTemp")
##get the phenology data
data  =  preparePhenologyData(dataFile=file, responseVariableName="DOY"
                              , onsetCovariateNames=vars, durationCovariateNames=vars
                              , taxonName="Sanguinaria_canadensis", removeOutliers=TRUE)
##run the Stan sampler
stanResult  =  runStanPhenology(type="full", responseData = data$responseData
                                , onsetCovariateData = data$onsetCovariateData
                                , durationCovariateData = data$durationCovariateData
                                , partitionDataForPriors = TRUE)
##summarize the Stan run
stanSummary  =  summarizePhenologyResults(stanRunResult = stanResult
                                          , taxonName = "Sanguinaria_canadensis"
                                          , standardLinearModel = TRUE)
##make posterior predictive graph
pp  =  makePosteriorPredictivePlot(stanResult = stanResult, responseData = data$responseData
                                   , targetCovariateName = "SpringMonthlyAverageTemp"
                                   , onsetCovariateData = data$onsetCovariateData
                                   , durationCovariateData = data$durationCovariateData)
##display the posterior predictive graph
print(pp)
}
}
