% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkSTAN.R
\name{makeSTANchecks}
\alias{makeSTANchecks}
\title{Check STAN and fix dependencies}
\usage{
makeSTANchecks()
}
\value{
Return TRUE or FALSE (using the "invisible()" function) when the system is ready (or not) to do analyses using STAN.
}
\description{
Function will conduct checks to verify if the system has the required dependencies to install and run STAN together with "phenoCollectR". The function will attempt to correct the dependencies and install STAN if necessary. The function is interactive and will prompt the user by printing information to the console.
}
