% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenologyInference.R
\name{PI.Ok1}
\alias{PI.Ok1}
\title{Title}
\usage{
PI.Ok1(
  N,
  mu_O,
  sigma_O,
  minResponse = 0,
  maxResponse = 365,
  alpha = 0.05,
  type = c("GP", "BB")
)
}
\arguments{
\item{N}{}

\item{mu_O}{}

\item{sigma_O}{}

\item{minResponse}{}

\item{maxResponse}{}

\item{alpha}{}

\item{type}{}
}
\description{
Title
}
