% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenologyVisualization.R
\name{posterior_predictive_graphic}
\alias{posterior_predictive_graphic}
\title{Title}
\usage{
posterior_predictive_graphic(
  observed_data,
  filtered_results,
  targetCovariateName,
  responseVariableName,
  legend = TRUE,
  timeSlice1,
  timeSlice2,
  adjustYLim = T,
  minResponse = 0,
  maxResponse = 365
)
}
\arguments{
\item{observed_data}{}

\item{filtered_results}{}

\item{targetCovariateName}{}

\item{responseVariableName}{}

\item{legend}{}

\item{timeSlice1}{}

\item{timeSlice2}{}

\item{adjustYLim}{}

\item{minResponse}{}

\item{maxResponse}{}
}
\description{
Title
}
