% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{preparePhenologyData}
\alias{preparePhenologyData}
\title{Prepare data for analysis}
\usage{
preparePhenologyData(
  dataFile,
  responseVariableName,
  onsetCovariateNames,
  durationCovariateNames,
  removeDuplicateRows = TRUE,
  removeOutliers = FALSE,
  removeIncomplete = TRUE,
  dataSummaryDirectory = NA,
  taxonName = NA,
  origN = NA
)
}
\arguments{
\item{dataFile}{A matrix object or a path to a file. See 'Details' for the correct configuration of this file.}

\item{responseVariableName}{The name of the response variable. Must match the name of a column in 'dataFile'.}

\item{onsetCovariateNames}{A vector with the names of the onset covariate variables.  Must match the name of a column in 'dataFile'.}

\item{durationCovariateNames}{A vector with the names of the duration covariate variables.  Must match the name of a column in 'dataFile'.}

\item{removeDuplicateRows}{If duplicated rows should be removed (TRUE or FALSE).}

\item{removeOutliers}{If outliers should be removed (TRUE or FALSE).}

\item{removeIncomplete}{If rows with missing data should be removed (TRUE or FALSE).}

\item{dataSummaryDirectory}{If a directory should be created to save the summary of the results.}

\item{taxonName}{The name of the taxon.}

\item{origN}{Explain what this is and the input options.}
}
\value{
Returns a list object with the elements necessary to use the "runStanPhenology" function. See 'Example".
}
\description{
This function prepares the phenological data for Bayesian estimation analysis using "runStanPhenology". The main inputs are the path to the data file and variable names. Note that all variable names should match the names of the columns in file.
}
\examples{
\donttest{
##get the file name with data for the blood root plant
file <- system.file("data", "Sanguinaria_canadensis.Full.txt", package = "phenoCollectR")
##get the phenology data
data <- preparePhenologyData(dataFile=file, responseVariableName="DOY", onsetCovariateNames=c("Year","Latitude","Elevation"), durationCovariateNames=c("Year","Latitude","Elevation"), taxonName="Mertensia_virginica", removeOutliers=TRUE)
##run the Stan sampler
stanResult <- runStanPhenology(type="full", responseData = data$responseData, onsetCovariateData = data$onsetCovariateData, durationCovariateData = data$durationCovariateData, partitionDataForPriors = TRUE)
##summarize the Stan run
stanSummary <- summarizePhenologyResults(stanRunResult = stanResult, taxonName = "Mertensia virginica",standardLinearModel = TRUE)
##make posterior predictive graph
pp <- makePosteriorPredictivePlot(stanResult = stanResult, responseData = data$responseData, targetCovariateName = "Year", onsetCovariateData = data$onsetCovariateData, durationCovariateData = data$durationCovariateData)
}
}
