% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{runStandardLinearModel}
\alias{runStandardLinearModel}
\title{Fit a standard linear model to the observed collection times, T}
\usage{
runStandardLinearModel(responseData, onsetCovariateData, durationCovariateData)
}
\arguments{
\item{responseData}{A vector of the response data.}

\item{onsetCovariateData}{A data frame with each named column providing a covariate for the onset model.}

\item{durationCovariateData}{A data frame with each named column providing a covariate for the duration model.}
}
\value{
The fit linear model object that is the output of the R lm function.
}
\description{
Fits a standard linear model to the observed collection times based on the covariates used to model the onset and duration.
}
\details{
Data objects for the input to this function can be obtained from the preparePhenologyData function and are provided with the output of the Stan run.

This function is run automatically when the standardLinearModel parameter of the summarizePhenologyResults function is set to TRUE, with corresponding results provided in the output data frame columns 'standard.linear.model'  and 'in.95CI'.
}
