% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenologyInference.R
\name{runStanPhenology}
\alias{runStanPhenology}
\title{Run a Bayesian analysis of phenological events using biocollection data (Core Package Function!)}
\usage{
runStanPhenology(
  type = c("intercept-only", "full", "multistage-full", "multistage-overlap-full"),
  responseData = NULL,
  stage = NULL,
  stageCounts = NULL,
  preN = NULL,
  visN = NULL,
  postN = NULL,
  nStages = NULL,
  hyperparams_noCovariates = NULL,
  onsetCovariateData = NULL,
  nOnsetCovariates = NULL,
  durationCovariateData = NULL,
  nDurationCovariates = NULL,
  onsetHyperBeta = NULL,
  onsetHyperBetaMean = NULL,
  onsetHyperBetaSD = NULL,
  onsetHyperAnchor = NULL,
  durationHyperBeta = NULL,
  durationHyperBetaMean = NULL,
  durationHyperBetaSD = NULL,
  durationHyperAnchor = NULL,
  sigmaHyper = NULL,
  totCounts = NULL,
  fixedCounts = FALSE,
  minResponse = 0,
  maxResponse = 365,
  scale = 50,
  maxDiv = 0,
  setStringent = TRUE,
  runMAP = FALSE,
  processExtremes = FALSE,
  N = NULL,
  partitionDataForPriors = FALSE,
  byPassChecks = FALSE,
  priorLevel = 2,
  debug = 0,
  nXs = 101,
  nReps = 100,
  calculatePPD = FALSE,
  bypass = FALSE,
  ...
)
}
\arguments{
\item{type}{The type of model to be implemented. Five options are available. Option 'full' will model mean onset and mean duration as a linear function of covariates for presence-only data, 'intercept-only' will model the distribution of onset times and the distribution of durations with no consideration of covariates for presence-only data. Option 'multistage-full' includes randomly sampled times of individuals, their phenological stage, and covariate data, with one linear model per stage, including for the interval between the start of a time period and the first stage in that time period. Option 'multistage-overlap-full' is the most general option. This option models individuals with 1 or more units that can occur in different stages. For example, an individual plant can have multiple floral units in different stages: bud, open flower, immature fruit, mature fruit. There is no multistage option for no covariates, but if you want to run such a model, set up a dummy covariate with a constant value and ignore the parameters for the linear model in the output. The Examples section below provides example usage of each scenario (default: intercept-only)}

\item{responseData}{A vector of the response data. The response data are simply the observed times of collection, often the day of year for biocollection data. This vector is prepared by the 'preparePhenologyData' function, or advanced users can prepare it using whatever tools they choose. The data should be in the original scale unless advanced users have a reason to use preprocessing transformations. (default: NULL)}

\item{stage}{Only for use with type 'multistage-full'. A vector of integers the same length as the number of observations. For each observation, the stage is the numeric value associated with each stage. The stage is 1 if the observation is in the interval between the start of the time period and before the first phenophase in the time period. (default: NULL)}

\item{stageCounts}{Only for use with type 'multistage-overlap-full'. A matrix of intergers with columns representing stages and rows representing individuals. Each row represents an individual, and the counts of units (e.g. floral units) in each stage corresponding to the columns is provided. If the stage is before a unit has developed or after units have senesced, provide a count for that stage indicating the average number of units per individual when the units are active (e.g., expected number of flowers for the plant). If the individual represents the unit and stages are sequential (non-overlapping within the unit), then there is one unit, and the counts should be 0 for all stages, except for the stage the individual is in, which should have a count of 1. For example, if each season, the individual produces a single flower and that flower is open, then the stage counts for pre-reproductive, in bud, in flower, in immature fruit, in fruit, post-reproductive would be: 0, 0, 1, 0, 0, 0. This will perform the identical analysis as type 'multistage-full'.}

\item{hyperparams_noCovariates}{For use with 'intercept-only' models. A vector of six elements that represent the mean and standard deviation (sd) hyperparameter values for the prior distributions of the mean onset, the mean duration, and the sigma parameters, in that order. Items in the vector can be named or unnamed. (default: NULL)}

\item{onsetCovariateData}{For use with '*full' models. A data frame with each column representing a covariate and each row representing a specimen observation. The number of rows must match the number of items in the responseData vector, and specimens are in the same order as they are in the response data vector. The function 'preparePhenologyData' can prepare this data frame, or advanced users can prepare the data frame using their own tools. Data should be in the original scale, or advanced users can transform the data as deemed appropriate. (default: NULL)}

\item{durationCovariateData}{For use with '*full' models. A data frame with each column representing a covariate and each row representing a specimen observation. The number of rows must match the number of items in the responseData vector, and specimens are in the same order as they are in the response data vector. In the 'multistage-full' model, the same covariates are used for onset and duration models. The function 'preparePhenologyData' can prepare this data frame, or advanced users can prepare the data frame using their own tools. Data should be in the original scale, or advanced users can transform the data as deemed appropriate. (default: NULL)}

\item{onsetHyperBeta}{For use with '*full' models. A data frame with two columns to provide hyperparameter values for the covariate slope coefficients for the onset model. The first column is the the mean hyperparameter, and the second is the standard deviation parameter. Each row corresponds with a separate covariate. The first row in the data frame corresponds to the covariate in the first column of the covariate data frame (onsetCovariateData), the second row with the second covariate, and so forth. If there are \eqn{K_O} columns in the covariate data frame, there should be \eqn{K_O} rows in this data frame. Can be left at default when priorLevel is set to 0, in which case Stan default priors are used (not recommended). (default: NULL)}

\item{onsetHyperAnchor}{For use with '*full' models. A two-element vector with the mean of the prior and the standard deviation of the prior for the onset anchor (the marginal mean onset time value). Can be left at default when priorLevel is set to 0, in which case Stan default priors are used (not recommended). (default: NULL)}

\item{durationHyperBeta}{For use with '*full' models. A data frame with two columns to provide hyperparameter values for the covariate slope coefficients for the duration model. The first column is the the mean hyperparameter, and the second is the standard deviation parameter, or for the 'multistage-full' model, a table with three columns, with the third indicating which stage the parameters are for. Each row corresponds with a separate covariate. The first row in the data frame corresponds to the covariate in the first column of the covariate data frame (durationCovariateData), the second row with the second covariate, and so forth. If there are \eqn{K_D} columns in the covariate data frame, there should be \eqn{K_D} rows in this data frame. Can be left at default when priorLevel is set to 0, in which case Stan default priors are used (not recommended). (default: NULL)}

\item{durationHyperAnchor}{For use with '*full' models. A two-element vector with the mean of the prior and the standard deviation of the prior for the duration anchor (the marginal mean phenophase duration value), or for the 'multistage-full' model, a data frame with two columns, the first being the mean and the second being the standard deviation, as before, but each row is for a different stage. Can be left at default when priorLevel is set to 0, in which case Stan default priors are used (not recommended). (default: NULL)}

\item{sigmaHyper}{For use with '*full' models. A two-element vector with the mean and standard deviation of the prior distribution for the sigma parameter, or for the 'multistage-full' model, a data frame with two columns, the first being the mean and the second being the standard deviation, as before, but each row is for a different stage. Sigma is the standard deviation of the onset distribution and is also the standard deviation of the cessation distribution. Can be left at default when priorLevel is set to 0, in which case Stan default priors are used (not recommended). (default: NULL)}

\item{minResponse}{Minimum value of the response (e.g., day of year); must be set to 0 (i.e., at default) under current implementation (default = 0)}

\item{maxResponse}{Maximum value of the response (e.g., day of year); typically 365 for Gregorian calendar (default = 365)}

\item{maxDiv}{The maximum number of divergences to be tolerated during the Stan posterior sampling. This should be 0 unless a biased sample from the posterior is acceptable. (default: 0)}

\item{setStringent}{Boolean flag to indicate more stringent sampling during Stan runs. Specifically, adapt_delta = 0.99, max_treedepth = 15. Setting to TRUE reduces the chances of divergences, but run time is slower. Usually the difference in run time is not of practical concern, so the default is fine. (default: TRUE)}

\item{runMAP}{Boolean flag to indicate if Stan should be used to estimate the maximum a posteriori (MAP) values. (default: TRUE)}

\item{processExtremes}{Boolean flag to indicate if parameters for the phenological extreme distributions should be estimated during the analysis. If set to TRUE, asymptotic estimates of the expected first onset and expected last cessation are sampled. Because asymptotic estimates are provided, estimates are made very quickly. A limitation is that an estimate of the population size is needed for this analysis. (default: TRUE)}

\item{N}{The population size. Needed when processExtremes is set to TRUE. (default: 500)}

\item{partitionDataForPriors}{Boolean flag to indicate if methods to automate the specification of prior hyperparameters should be used. In particular, 30\% of the data are partitioned into a set to estimate the hyperparameters, and 70\% of the data are used for the Bayesian analysis using Stan. If set to TRUE, all other user-provided hyperparameter values (e.g., hyperparams_noCovariates, onsetHyperBeta, onsetHyperAnchor, etc.) are ignored. This is \emph{not} recommended if accurate estimates of duration and onset parameters are needed, but estimates of covariate coefficients tend to be accurate. (default: FALSE)}

\item{byPassChecks}{Boolean flag indicating whether to bypass checks for Stan. If Stan is not located, an attempt will be made to install Stan. Recommended to keep at default. (default: FALSE)}

\item{priorLevel}{Specifies types of priors by an integer. Setting to 0 will use flat priors. Above 0 will use the user-input prior hyperparameters with normally distributed priors. For presence-only data, leave at default. (default: 2)}

\item{nXs}{For use with 'multistage-full' analyses only. Number of increments along the x-axis for posterior predictive plotting. (default: 101)}

\item{nReps}{For use with 'multistage-full' analysis only. Number of replicates per x-axis increment to sample for posterior predictive estimates. (default: 100)}

\item{calculatePPD}{For use with 'multistage-full' analyses only. Boolean indicating whether the posterior predictive distribution should be estimated. (default: FALSE)}

\item{...}{Parameters to be input into the Stan sample function. Not currently implemented. Do not use.}

\item{maximizeSampleSize}{Boolean flag for use when partitionDataForPriors is set to TRUE. If a user wants 100\% of the data to be used for the Bayesian analysis, set this to TRUE. This is statistically invalid because the same data are used to estimate the prior hyperparameters and carry out the full Bayesian analysis. When the sample size is very small, and no prior information is available, the error on the estimates may be unacceptably large unless the fuller dataset is used for inference. Deprecated and no longer implemented. (default: FALSE)}
}
\value{
Details of what is returned will depend on which 'type' of model is used. See 'type' parameter.

In the case of \bold{'*full' type}, a list with the following items is returned:

data: the data that were passed to Stan. These include the scaled and translated response and covariate data.

responseData: the original response data

onsetCovariateData: the original covariate data for the onset model

durationCovariateData: the original covariate data for the duration model

onsetHyperBeta: the mean and standard deviation hyperparameter values for the onset model covariate coefficients

onsetHyperAnchor: the mean and standard deviation hyperparameter values for the onset anchor (marginal mean onset)

durationHyperBeta: the mean and standard deviation hyperparameter values for the duration model covariate coefficients

durationHyperAnchor: the mean and standard deviation hyperparameter values for the duration anchor (marginal mean duration)

sigmaHyper: the mean and standard deviation hyperparameter values for sigma

result: the sample from the Stan sampling

model: the Stan model

maxDiv: the maximum allowed number of divergences for the run

minResponse: the minimum possible response value

maxResponse: the maximum possible response value

setStringent: whether setStringent was set to TRUE.

error: whether an error was caught during the Stan run.

error_m: an error message, possibly indicating no error occurred.

In the case of \bold{'intercept-only'} type, a list with the following items is returned:

data: the data that were passed to Stan. These include the scaled and translated response and covariate data.

responseData: the original response data

hyperparameters: the hyperparameters used by the Stan model

sample: the sample from Stan

model: the Stan model

runMap: Boolean indicates of MAP estimate was made

MAP: Stan's MAP estimate, if selected. NA if runMap is FALSE

N: input population size

minResponse: the minimum possible response value

maxResponse: the maximum possible response value

setStringent: whether setStringent was set to TRUE.

maxDiv: the maximum allowed number of divergences for the run

error: whether an error was caught during the Stan run.

error_m: an error message, possibly indicating no error occurred.
}
\description{
This function is the core of the phenoCollectR package. It wraps many other functions to make the Bayesian analysis of phenological events using biocollection data as streamlined as possible, yet flexible if need be. In particular, it aims to connect variation in life history timing to environmental variation.

Models with or without covariates can be used. Covariates can be climate variables, physical environment variables, physiological states, genetic states, and so forth - anything that may contribute to phenological variation. However, as currently implemented, only continuous covariates are supported.

Mean phenophase onset time and mean phenophase duration are modeled as linear functions of the covariates, if supplied. Otherwise, just the mean phenophase onset time and the mean phenophase duration are modeled without covariates. Separate covariates can be used to model the mean onset time and the mean phenophase duration. However, some downstream methods for summaries and plotting expect the covariates to be the same for both models.

The helper function 'preparePhenologyData' can be used to prepare the response data and the covariate data for input to this function.

The theory developed by Hearn et al. (XXXX) unifies the models for onset and duration to derive the theoretical distribution of observed collection times. Bayesian inference is carried out using the probabilistic programming language Stan, which draws samples from the posterior distribution of the model parameters. These samples can be used for downstream analyses. For example, to estimate the marginal posterior probability that a slope coefficient associated with a covariate is negative, one simply counts the number of posterior draws where the coefficient is negative and divides by the total number of draws from the posterior distribution.

As part of a Bayesian analysis, and to reduce model identifiability issues with presence-only data, a prior distribution is needed for the analysis. Flat priors will not work for this analysis. Because the setting of priors can be challenging, this function can optionally implement methods that will automate the specification of priors for you. Advanced users can manually tailor prior hyperparameters and achieve weakly informative priors, if desired. Under the current implementation, the functional form of the priors is pre-set.

Users can also optionally choose to analyze phenological extremes. To do so requires an estimate of the population size from which biocollection records were obtained. The concept of "population" in this case can be quite loose and can be the metapopulation consisting of all populations from which samples were selected.

The parameters to be estimated for the 'full' model include the marginal mean onset time (\eqn{\mu_O}), the marginal mean duration (\eqn{\mu_D}), one parameter (\eqn{\beta_O} coefficient) for each covariate used to model the mean onset time, one parameter (\eqn{\beta_D} coefficient) for each covariate used to model the mean duration time, and one \eqn{\sigma_O} parameter which models both the standard deviation of the distribution of onset times and the standard deviation of the distribution of durations. If there are \eqn{K_O} onset covariates, and \eqn{K_D} duration covariates, then there is a total of 3 + \eqn{K_O} + \eqn{K_D} free parameters (including the two \eqn{\alpha} intercept terms and \eqn{\sigma_O}).

The parameters to be estimated for the 'intercept-only' model include the marginal mean onset time (\eqn{\mu_O}), the marginal mean duration (\eqn{\mu_D}), and the sigma parameter (\eqn{\sigma_O}).

Other parameters associated with phenophase cessation events and phenological extremes are deterministic functions of the free parameters based on the implemented Gaussian process model.

A summary of the Stan run is available through the helper function 'summarizePhenologyResults', or advanced users can access the Stan sample directly from the output of this function for customized downstream analyses using other packages designed for the analysis of samples from posterior distributions, such as 'posterior' and 'bayesplot'.

\bold{\emph{BACKGROUND FRAMEWORK AND THEORY}}

\bold{DATA TYPES}

There are two main types of datasets that can be analyzed. When only observation times of individuals in the stage of interest are available, these are presence-only data (PO). When observation times are available for multiple stages of interest, these are multistage data. Theory for these two types of datasets differs.

The following sections can safely be skipped if not interested in theory.

\bold{NOTATION}

Both visible and latent events are modeled with random variables (RVs).

RVs are symbolized with the first letter of the state of interest: Onset (O), Duration (D), Cessation (C), observed collection Times (T).

RVs representing extremes are subscripted. For example, first onset in a population is \eqn{O_1}, and last cessation is \eqn{C_N} where \eqn{N} is the population size.

Parameters are subscripted by the associated random variable. For example, mean duration is \eqn{\mu_D} and the standard deviation of the last cessation is \eqn{\sigma_{C_N}}. In the linear models, intercept parameters are represented by \eqn{\alpha} and slope coefficients are \eqn{\beta}. So, for the model of mean onset, the intercept is \eqn{\alpha_O} and a slope coefficient for the \eqn{i^{th}} covariate is \eqn{\beta_{O,i}}.

The model implemented in Stan is based on the properties of a degenerate Gaussian process that samples lines with intercepts whose values are normally distributed. When a line passes a first threshold at \eqn{\mu_O}, the process begins, and when that line passes a second threshold at \eqn{\mu_D}, the process ends. This process results in normally-distributed onsets: \eqn{O \sim N(\mu_O, \sigma^2_O)}, a constant duration \eqn{D}, and a normally-distribed cessation: \eqn{C \sim N(\mu_O + D, \sigma^2_O)}.

The process of interest occurs in a time period (TP), such as a year for cyclical annual processes, or a day for cyclical diel processes, that has a minimum time of \eqn{m=0} and a maximum time of \eqn{M}. For a year, the day of year time is between \eqn{m=0} and \eqn{M=365}.

The dataset consists of a population of individuals undergoing the process of interest contained within the TP. Onset and duration times can vary for each individual in the population.

Onset times have intrinsic noise due to \eqn{\sigma_O}, whereas durations do not. Both onset times and duration lengths can vary in relation to internal states such as energy store level or organismal size, or to external conditions such as climate variation (temperature summaries for spring, etc.) or environmental variation (soil type, moisture levels, light levels, etc.). Any number of covariates can be used with standard cautions about overly complex models (overfitting, model identifiability, parameter uncertainty, collinearity, etc.).

The sample design is as follows. Individuals are randomly sampled from a population at random points in time during the time period.

The only RVs with directly observed states are the observed times (\eqn{T}) and the observed phenological stages (\eqn{S}). All the other RVs are latent (hidden) and estimated from the observed times and associated stages.

For the theory, we simplify things by examining a time period (TP) starting at time 0 and of length 1 time unit. Data sampled from other length time periods can be translated to start at time 0 by subtracting \eqn{m} from all times and scaled to a length of 1 by dividing all sampled times by \eqn{M-m}. In what follows, we assume such transformations have already been made. These transformations are handled automatically by this runStanPhenology function. If users wish to transform their data in other ways prior to analysis, we leave this up to the user.

\bold{THEORY: MULTISTAGE}

For \bold{multistage} data, all observed times are kept and the stage (e.g., flowering, fruiting, active growth, dormancy) is recorded with an integer label.
For the core theory, we need to determine the probabilities of each stage given the sampled time and model parameters. The core parameters are the mean onset of the first full stage in the TP, the mean durations, and the variation in onset times.

Here, the notation is slightly different that that presented above. \eqn{O_1} represents the onset time of the first stage, rather than the first onset phenological extreme.

If you are interested in doing an analysis of data before a phenophase of interest, during a phenophase of interest, and after a phenophase of interest (BDA analysis), code the data with the following stages: 1 = before, 2 = during, and 3 = after.

The probability that an individual is in a stage that is before the first full stage begins (ie, stage 1) is \deqn{P_{1|j}(O_1>t_j|\theta)=1-\Phi(z_{1|j})} where \eqn{t_j} is the observed time of collection of the \eqn{j^{th}} individual, \eqn{\Phi(z) \equiv P(Z \le z)} is the cumulative distribution function of a standard normal random variable, and z values are the standardized times \eqn{z = \frac{t-\mu}{\sigma}}.

The probability that an individual \eqn{j} is in stage \eqn{s} is the probability that the sampled time is after the onset of stage \eqn{s} and before the onset of the next stage \eqn{s+1}. Using De Morgan's law for probabilities, \deqn{P_{s|j}(O_s \leq t_j, O_{s+1} > t_j|\theta) = \Phi(z_{s|j})-\Phi(z_{s+1|j})}

Last, the probability that an individual is in the last stage is the probability that the sampled time is after the onset of the last stage S: \deqn{P_{S|j}(O_S \leq t_j|\theta) = \Phi(z_{S|j})}

The likelihood follows a categorical random variable: \deqn{L(\theta) = \prod_{j=1}^{j=N}\prod_{s=1}^{s=S}P_{s|j}^{I(s_j = s)}} where the indicator function \eqn{I(s_j=s)} returns 1 whenever the stage for the \eqn{j^{th}} individual equals \eqn{s} and returns 0 otherwise.

\bold{THEORY: OVERLAPPING MULTISTAGE}

Please note that implementation of the overlapping multistage model is in progress.  You may run the model and expect fairly accurate estimates, but the uncertainty estimates are unlikely to be accurate. Please use with discretion.

The multistage theory above is a special case of a more general multistage theory in which each individual produces one or more phenological units that can be at different developmental stages on the same individual. For example, a plant may produce multiple floral units in a season. In this general case, the categorical model for the single individual in one of multiple stages generalizes to a multinomial model where counts are provided for each stage.

For \bold{individuals with multiple units of organization each with multiple stages}, such as an individual plant with multiple floral units with stages 'bud', 'flower', 'immature fruit', and 'mature fruit', the counts of units in each stage are provided as input along with the time the individual was observed.

The probability that a unit is in a particular stage \eqn{s} given the time of observation, \eqn{P_{s|j}}, are the same probabilities as for the multistage theory above.

In principle, there are many options to model the distribution of stages of units within an individual at a particular time of observation. One could model the developmental dynamics of the units and use this model to generate the distribution of stages. This is ideal, but requires detailed knowledge of the developmental dynamics of individuals. As an idealization, here, the probability mass function of the stages of units within an individual is modeled as multinomial, which models each unit as developing independently of the other units within the individual once the unit itself has begun development (The units are synchronized with noise within an individual). The stage probabilities \eqn{P_{s|j}} provide the probabilities of occuring in each stage for the multinomial model.

We noted that this overlapping multistage model is a generalization of the multistage model presented above. However, there are stages when units are not yet present (pre) or when units have abscised (post) from the individual. In such cases, it is impossible to count the number of units in these latent pre and post stages. Additional information in the way of total expected units produced by the individual is therefore needed when latent pre and post stages are present.

As a generalization of the multistage model, a special case of the overlapping multistage model reduces to the multistage model. When an individual is the single unit of phenological activity, then a 0 is scored for all stages that the individual is not in, and a 1 is scored for the stage the individual is in. In this context, the multinomial probability reduces to the categorical probabilties defined for the multistage analysis above.

\bold{THEORY: PRESENCE-ONLY (PO)}

For \bold{presence-only (PO)} data, only the observed times \emph{during} the process of interest are recorded and all other times are discarded.

The theory for presence only is a bit more complicated than the theory for multistage data because marginalization is required to normalize the probability of being in a stage. The ultimate goal for PO analysis is to infer the parameters that describe the distribution of observed sampled times. The other distributions of interest, including extremes, onsets, peak, and cessation, are functions of these parameters (with a caveat for extremes).

The distribution whose parameters we want to infer is the probability density of the observed times given the individual was sampled during the phenological stage of interest (stage \eqn{s_j=s}) and given the parameters of the onset distribution and duration: \deqn{p_{j|s} \equiv p(t_j|s_j=s,D,\mu_O,\sigma_O)} For simplicity, let \eqn{\theta \equiv (D,\mu_O,\sigma_O)}. Using Bayes rule, \deqn{p(t_j|s_j=s,\theta) = \frac{P(s_j=s|t_j,\theta)p(t_j|\theta)}{P(s_j=s|\theta)}}

The probability that a sampled individual \eqn{j} is in the focal stage of interest, \eqn{s}, given the sampled time and parameters for individual \eqn{j}, was defined above as \eqn{P_{s|j}}.

The sampling strategy defines \eqn{p(t_j|\theta)}. In the current implementation, the sample of times is assumed to be a random sample and independent of \eqn{\theta}, i.e., \eqn{p(t_j|\theta)=\frac{1}{M-m}}. Since all observed times are scaled to a TP of length 1 already, \eqn{p(t_j|\theta)=1}.

Finally, \eqn{P(s_j=s|\theta)} is the expected length of the stage of interest when the TP is length 1 and the stage occurs entirely within the TP. By the law of total probability, this is \deqn{\int_0^1 P(s_j=s,t_j|\theta)dt_j=\int_0^1 P(s_j=s|t_j,\theta)p(t_j|\theta)dt_j=\int_0^1 P(s_j=s|t_j,\theta)dt_j=\int_0^1P_{s|j}dt_j=\int_0^1 (\Phi(\frac{t-\mu_O}{\sigma_O})-\Phi(\frac{t-D-\mu_O}{\sigma_O}))dt_j} Hearn et al. (XXXX) in their Supporting Information Notes S1 provide a solution for this integral in terms of \eqn{\Phi} and \eqn{\phi} (the density function for the standard normal).

Defining the normalization constant, \eqn{c_j}, as \eqn{\frac{P(s_j=s|\theta)}{p(t_j|\theta)} = P(s_j=s|\theta)}, then the likelihood function for presence-only data is: \deqn{\prod_{j=1}^{j=N} \frac{p_{j|s}}{c_j}} In this case, the normalization constant is different for each individual once predictors for mean onset and duration are defined in terms of covariates, as each individual has different covariate values. So, it doesn't cancel. The next section develops this aspect of the model.

\bold{THEORY: MODELING COVARIATES}

In the current implementation, the mean onset time and the duration times are modeled as linear functions of covariates. Different covariates can be used for the onset model and for the duration model(s), but for the current implementation of multistage inference, the covariates must be the same for onset and duration model(s). If we let \eqn{\vec{X}} represent the vector of covariates and \eqn{\vec{\beta}} represent the vector of coefficients, then the following are the models for mean onset and mean duration: \deqn{O \sim N(\alpha_O + \vec{\beta_O} \cdot \vec{X}, \sigma_O^2)} and \deqn{D \equiv \mu_D = \alpha_D+\vec{\beta_D} \cdot \vec{X}} For multistage inference, there are multiple duration models, one for each stage, each with a different intercept \eqn{\alpha} and coefficients \eqn{\vec{\beta}}.

These models of mean onset and duration make the model hierarchical, so that with the inclusion of covariates, everywhere in the likelihood function where \eqn{\mu_O} or \eqn{D} appear, they are replaced by the above linear functions. If there are \eqn{K_O} covariates for the onset model and \eqn{K_D} (\eqn{K_D} = \eqn{K_O} for multistage models) covariates for the duration model(s), along with the two \eqn{\alpha} intercepts and the \eqn{\sigma_O} standard deviation parameter, the fully specified model has \eqn{2 + K_O + K_D + 1} estimable parameters, or S + S*K_O + 1 parameters for the multistage model.

\bold{THEORY: BAYESIAN ANALYSIS AND PRIORS}

In a Bayesian analysis, the probability density of the parameters given the data (which is called the posterior distribution) is proportional to the product of the likelihood (which is the probability of the data given the parameters) and the prior distribution (which is the marginal probability of the parameters): \eqn{Posterior \propto Likehood * Prior}.

The above sections provided derivations of the likelihood functions for both PO datasets and multistage datasets, so the only remaining thing needed to carry out a Bayesian analysis is to specify the prior distributions. The priors are defined for the \eqn{D,\beta_O,\alpha_O, \beta_D, \alpha_D, \sigma_O} parameters. In the current implementation, the priors can be flat, or they can be normally distributed.

Set the \emph{priorLevel} function parameter (described below) to 0 for a flat prior. Identifiability for PO analyses necessitates an informative prior.  Priors for the multistage model are automatically set to weakly informative priors if none are provided by the user.

Set the \emph{priorLevel} to any value greater than or equal to 1 for normally distributed priors. A normally distributed prior requires two hyperparameters for each parameter: the mean of the parameter and the standard deviation of the parameter. Users can manually input these hyperparameter values or can use automated prior specification.

Normal priors are used primarily for convenience and interpretability. Under standard regularity conditions, the Bernstein–von Mises theorem implies that as sample size increases, the posterior distribution becomes approximately multivariate normal, centered near the maximum likelihood estimate, with covariance given by the inverse Fisher information scaled by \eqn{1/n}. In this large-sample regime, posterior uncertainty is therefore well summarized by normal approximations and depends only weakly on the particular prior, provided the prior assigns positive density in a neighborhood of the true parameter value. This asymptotic behavior makes normal priors a natural default choice in many settings. Moreover, in sequential Bayesian analyses where researchers use a previous large-sample posterior as the prior for a subsequent analysis, the inherited prior will itself be approximately normal.
}
\examples{
\donttest{
##########################################################################################################################################################
##Run an example with EMPIRICAL COLLECTION TIMES of individuals in the phenophase (PRESENCE ONLY - PO) - type 'full'.

##DEFAULT PRIORS are used and are only reliable for slope coefficient estimations. 
#!!*Use a prior based on biological data for the duration if accurate duration estimation is a priority.*!!
#
#Get the FILE NAME with data for the blood root plant. Data files for 12 other species are also available
file  =  getDatasetPath("Sanguinaria_canadensis")
#
## ACCESS MORE SPECIES
help(getDatasetPath)
#
#COVARIATE NAMES - remove up to all but 1
vars = c("Latitude", "Year", "Elevation", "AnnualMonthlyAverageTemp"
         , "SpringMonthlyAverageTemp", "FirstQuarterMonthlyAverageTemp")
#
#EXTRACT PHENOLOGY DATA
data  =  preparePhenologyData(dataFile=file, responseVariableName="DOY"
                              , onsetCovariateNames=vars, durationCovariateNames=vars
                              , taxonName="Sanguinaria_canadensis", removeOutliers=TRUE)

\bold{#EMPIRICAL DATA STAN RUN WITH COVARIATES}
stanResult  =  runStanPhenology(type="full", responseData = data$responseData
                                , onsetCovariateData = data$onsetCovariateData
                                , durationCovariateData = data$durationCovariateData
                                , partitionDataForPriors = TRUE)
#
#SUMMARIZE 
stanSummary  =  summarizePhenologyResults(stanRunResult = stanResult
                                          , taxonName = "Sanguinaria_canadensis"
                                          )
#WRITE EMPIRICAL RESULTS
write.csv(stanSummary,"SCanadensis.Empirical.wCovariates.csv")
##
##
##########################################################################################################################################################
##Run more complicated examples, this time with SIMULATED PHENOLOGY and SIMULATED COVARIATES
#
#COVARIATE MODEL
covariate_namesOnset = c("v1", "v2")	
covariate_namesDuration = c("v1", "v3")	#Note that for presence-only data for one stage, different covariates can be used
cov_all = union(covariate_namesOnset, covariate_namesDuration)
#covariance matrix must have all variables (v1, v2, and v3 in this case)
covarianceMatrix = matrix(c( 1.0, 0.5, 0.3,
				      0.5, 2.0, 0.4,
				      0.3, 0.4, 1.5), nrow = 3, byrow = TRUE)
rownames(covarianceMatrix) = cov_all
colnames(covarianceMatrix) = cov_all
covariateMeans = c(50,60,80)
names(covariateMeans) = cov_all
#
#ONSET PARAMETERS
slopesOnset = c(1,-2)
names(slopesOnset) = covariate_namesOnset
mean_responseOnset = 150
noiseOnset = 3
#
#DURATION PARAMETERS
slopesDuration = c(3,1)
names(slopesDuration) = covariate_namesDuration
mean_responseDuration = 30
#
#SAMPLE SIZE
n=3000
#
#SIMULATE PHENOLOGY TIMES, OBSERVATION TIMES, AND COVARIATE DATA
#	The function simulatePopulationLatentIntervalStates can handle different covariates for onset and for duration of the stage of interest
simulated_data = simulatePopulationLatentIntervalStates(n=n,
                       betaOnset=slopesOnset, betaDuration=slopesDuration,
               covariateNamesOnset=covariate_namesOnset, covariateNamesDuration=covariate_namesDuration,
               covariateMeans = covariateMeans,
               covarianceMatrix = covarianceMatrix,
               meanOnset = mean_responseOnset, meanDuration = mean_responseDuration,
               sigmaOnset = noiseOnset)
#
#_________________________________________________________________________________________________________________________________________________________
#
#HYPERPARAMETERS - Note these are unbiased. Consider changing them from the true values set above
onsetHyperBeta = data.frame(slopesOnset,c(1,1))
onsetHyperAnchor = c(150, 5)
durationHyperBeta = data.frame(slopesDuration,c(1,1))
durationHyperAnchor = c(30, 3)
sigmaHyper = c(3,1)
#
#_________________________________________________________________________________________________________________________________________________________
#
#PO DATA - Extract data from one stage only
responseData = simulated_data$observedTime[simulated_data$stage==2]
onsetCovariateData = data.frame(simulated_data$v1[simulated_data$stage==2],simulated_data$v2[simulated_data$stage==2])
durationCovariateData = data.frame(simulated_data$v1[simulated_data$stage==2],simulated_data$v3[simulated_data$stage==2])
colnames( onsetCovariateData ) = covariate_namesOnset
colnames( durationCovariateData ) = covariate_namesDuration
#
\bold{#PO STAN RUN - type 'full'}
stanResultPO = runStanPhenology(
			type="full",
			responseData=responseData,
			onsetCovariateData=onsetCovariateData, durationCovariateData=durationCovariateData,
			onsetHyperBeta=onsetHyperBeta,onsetHyperAnchor=onsetHyperAnchor,
			durationHyperBeta=durationHyperBeta,durationHyperAnchor=durationHyperAnchor,
			sigmaHyper=sigmaHyper,
			minResponse=0,maxResponse=365,
			runMAP=TRUE,processExtremes=TRUE,N=500,
			partitionDataForPriors=FALSE,
			maxDiv=0,setStringent=TRUE,
			priorLevel=2
			)
##summarize the Stan run
stanSummaryPO  =  summarizePhenologyResults(stanRunResult = stanResultPO
                                          , taxonName = "Simulated"
                                          )
##write results to file
write.csv(stanSummaryPO,"Simulated.PresenceOnly.informativePriors.csv")
##
##
##########################################################################################################################################################
##Run a full multistage analysis with SIMULATION PARAMETERS themselves simulated and SIMULATED COVARIATES with MULTIPLE STAGES
#
library(dplyr)
library(posterior)
library(ggplot2)
 #Set the parameters 
 nStages = 4         #Number of stages to be simulated (will be 1 more than this once "unrolled" from cyclical time period
 nCovariates = 3     #Number of covariates to be simulated
 n = 500             #Sample size
 minResponse = 0 		#The minimum observed time should always be at least 0
 maxResponse = 365		#Days in the year for yearly cycles

 #Automated - many options available for more control
 simulatedData = simulateMultistageData(n=n, nStages=nStages,nCovariates=nCovariates)
 
 #Plotting 
 	#Set which covariate is the x-axis in the plot
 targetCovariateIndex = 1
 
     #Set colors
 stageColors = RColorBrewer::brewer.pal(nStages+1, "Set1")

 #Plot the sampled data color-coded by stage 
 trueModels = plotMultistageSimulation(simulatedData=simulatedData,
                  targetCovariateIndex=targetCovariateIndex,
                  stageColors=stageColors,
                  drawLatentOnset=FALSE,
                  drawObserved=TRUE,
                  drawTrueModel=FALSE,
                  drawInferredModel=FALSE,
                  shadeStage=TRUE,
                  minResponse=minResponse, maxResponse=maxResponse)
 
 #Run inference with Stan - THIS MAY TAKE A WHILE!
 stanResults =runStanPhenology(
     type="multistage-full",                             #model with many stages and covariates
     responseData=simulatedData$outputData$sampledTime,  #observed collection times
     stage=simulatedData$outputData$sampledStage,        #observed stage at collection time
    nStages=nStages,				            #number of stages - needed, as not all stages are necessarily sampled
     onsetCovariateData=simulatedData$X,                 #covariate data (same as for duration)
     maxDiv=4000,					    #should be set to 0, but in case of the stray divergence, set high for this example
     calculatePPD=TRUE
 )

#Plot the posterior predictive distribution - THIS TAKES SOME TIME
#BLACK dotted lines = true marginal mean onsets (marginalized across other covariates)
#COLORED solid lines = posterior predictive mean onset (no marginalizing across other covariates)
#RED dotted lines = mean posterior marginal mean onsets (marginalized across other covariates)
p = makeMultistagePosteriorPredictivePlot(stanResult=stanResults,
   responseData=simulatedData$outputData$sampledTime,
   targetCovariateName="cov1",
   covariateData=simulatedData$X,
   stageData=simulatedData$outputData$sampledStage,
   nStages=nStages,    #with stage coding, an extra stage is added
   y_pred=TRUE)          #the predicted responses are calculated when calculatePPD is set to TRUE during Stan run (above)
 
 #Plot the true mean stage boundaries (based on used simulation parameters) with black dotted lines
 for(i in 1:(nStages-1)) {
   p = p + geom_abline(intercept = trueModels$trueIntercepts[i], slope = trueModels$trueSlopes[i], color = "black", linewidth = 0.5, linetype="dashed")
  }

 #Extract basic summary data
 probs = c(2.5, 97.5)
 measures=c("mean", "median", "sd", "mad")
 
 summary = print(
   stanResults$result$summary(
     variables = c(
       "sigma","anchor_d", "beta_d", "alpha_d", "anchor_o", "beta_o", "alpha_o"
      ),
     quantiles = ~ quantile2(., probs=probs/100),
     measures
   ),
   n = Inf
 ) \%>\% as.data.frame()
 
 
  #Set up vectors to store mean durations (means) and mean onsets (means_o) with low and high bounds of credible intervals.
 
  #Extract slopes and intercepts (anchors) and overlay the inferred lines and 95\% BCI's onto earlier plot as colored lines and shaded regions
  for(j in 1:(nStages)) {
          beta = rep(0,nCovariates)
          beta_low = rep(0,nCovariates)
          beta_high = rep(0,nCovariates)
          varO = paste0("anchor_o[", j, "]")
          alpha = summary[summary$variable==varO, "mean"]
          alpha_low = summary[summary$variable==varO, paste0("q",probs[1])]
          alpha_high = summary[summary$variable==varO, paste0("q",probs[2])]
 
         for(i in 1:nCovariates) {
                  varO = paste0("beta_o[", j, ",", i, "]")
                  beta[i] = summary[summary$variable==varO, "mean"]
                  beta_low[i] = summary[summary$variable==varO, paste0("q",probs[1])]
                  beta_high[i] = summary[summary$variable==varO, paste0("q",probs[2])]
          }
         model = phenoCollectR:::true_marginal_line(alpha=alpha, beta=beta, mu=simulatedData$covariateMeans, Sigma=simulatedData$Sigma, j=targetCovariateIndex)
         model_low = phenoCollectR:::true_marginal_line(alpha=alpha_low, beta=beta_low, mu=simulatedData$covariateMeans, Sigma=simulatedData$Sigma, j=targetCovariateIndex)
         model_high = phenoCollectR:::true_marginal_line(alpha=alpha_high, beta=beta_high, mu=simulatedData$covariateMeans, Sigma=simulatedData$Sigma, j=targetCovariateIndex)
   p = p + geom_abline(intercept = model$intercept, slope = model$slope, color = "red", linewidth = 0.5, linetype="dashed")
  }

 print(p)
 
 
##
##
##
##########################################################################################################################################################
##Conduct an analysis of individuals with multiple phenological units and each unit having one of multiple possible stages.
#(For example: an individual plant with multiple flower units, each in a different stage (budding, flowering, fruiting) with pre-reproductive and post-reproductive stages possible with no flowers. 
###SIMULATION PARAMETERS
 n = 500            #individuals sampled
 meanFlowers = 20   #mean number of flowers per individual

 #There can be at most 1 stage before units are visible (pre) or 1 stage after units have abscised, but not both. If there are no stages with visible units, then there can be multiple pre stages only. 
 nPre = 1              #number of stages before visible units develop
 nVis = 2              #number of stages with visible units
 nPost = 0             #number of stages after visible units

 nCovariates = 2     #number of simulated environmental factors
 nStages = nPre+nVis+nPost     #total number of stages: due to a bug, this must be 3 or greater

 maxResponse = 365   #maximum day of year of observation
 minResponse = 0     #minimum day of year of observation - must stay at 0

 scale = 50    #for transforming the scale within Stan (should be around 10 to 50 - default is fine)

 meanOnsetSpread = sample(10:20, size=nStages, replace=TRUE)       #mean weight for each stage duration (Dirichlet)

   minStageVariance = (1/10)*(maxResponse*(10/sum(meanOnsetSpread)))^2  #set smallest simulation stage onset standard deviation to expected smallest duration - plenty of overlap!
 maxStageVariance = 2*minStageVariance

 # Simulate the data
 sim = phenoCollectR:::simulateMultistageOverlapData(n=n, meanUnits=meanFlowers, nPre=nPre, nVis=nVis, nPost=nPost, nCovariates=nCovariates, minStageVariance = minStageVariance, maxStageVariance=maxStageVariance, meanOnsetSpread=meanOnsetSpread)

 # Plot the simulated data
 par(mfrow = c(1, 2))  # 1 row, 2 columns
 plotMultistageSimulation(simulatedData = sim, includeOverlap=TRUE, drawInferredModel=FALSE, main="X: times of observation; solid lines: true model of mean onset times", drawLatentOnset=FALSE)

 #Get the unit count matrix (e.g., number of floral units in floral stages)
 stageCountMatrix <- t(
 sapply(sim$simulatedIndividuals, `[[`, "stageCounts")
 )

 #Get the true total counts for each individual (including missing ones in pre or post)
 #Currently treated as data (no prior)
 #Not needed if all stages are stages that produce visible units (vis stages)
 trueTotCounts = sapply(sim$simulatedIndividuals, `[[`, "nUnits")

 #Get the observed times
 t_obs <- sapply(sim$simulatedIndividuals, `[[`, "sampledTime")

 #Print simulated data
 print(cbind(stageCountMatrix,t_obs))

 stanResults =runStanPhenology(
   type="multistage-overlap-full",       #model with many stages and covariates
   responseData=t_obs,  #observed collection times
   stageCounts=stageCountMatrix,        #observed stage counts at collection time
   preN = nPre,
   visN = nVis,
   postN = nPost,
   onsetCovariateData=sim$simulatedData$X,           #covariate data (same as for duration)
   totCounts = trueTotCounts,  #currently using true simulated times
   fixedCounts = TRUE,		#not used
   minResponse = minResponse,
   maxResponse = maxResponse,
   scale = scale,	    	#scale for transforming data in Stan - default is usually fine
   bypass=T,			#force running of multistage overlap, acknowledging still in development
   maxDiv=4000             #should be set to 0, but in case of the stray Stan divergence, set high for this example
 )

         #Extract basic summary data
         stanSummary = phenoCollectR:::getMultistageSummary(stanResult=stanResults)
         nDivergent = sum(stanResults$result$diagnostic_summary()$num_divergent)
         nMaxTreeDepth = sum(stanResults$result$diagnostic_summary()$num_max_treedepth)

         stageColors = rainbow(nStages)
         trueModels = plotMultistageSimulation(simulatedData=sim, stageColors=stageColors, includeOverlap=TRUE, shadeStage=FALSE, drawLatentOnset=FALSE, drawTrueModel=FALSE, drawInferredModel=FALSE, main="Dotted red: true model; Dotted black: mean posterior estimate")

         targetCovariateIndex = 1
         # Add estimated mean stage boundaries
    #Black dotted lines are the mean posterior estimates
    #Red dotted lines are the true model lines
         for(i in 2:nStages) {
           model = phenoCollectR:::true_marginal_line(alpha=stanSummary$mean_o[i-1], beta=stanSummary$beta_o[i-1,], mu=sim$simulatedData$covariateMeans, Sigma=sim$simulatedData$Sigma, j=targetCovariateIndex)
           abline(a=model$intercept, b=model$slope, col="black", lwd=3, lty="dotted")
           abline(a=trueModels$trueIntercepts[i-1], b=trueModels$trueSlopes[i-1], col="red", lwd=3, lty="dotted")
         }
    print("True models (a = intercepts / stage durations / anchors, b = slopes)")
    print(trueModels)
##
##
##########################################################################################################################################################
##Conduct an analysis with NO COVARIATES (NC) using SIMULATED DATA
#
#SIMULATION PARAMETERS 
mean_responseOnset = 150
sdOnset = 3
mean_responseDuration = 30
#
#SAMPLE SIZE
n=5000
#
#SIMULATE
simulated_data = simulatePopulationLatentIntervalStates(n=n,
               meanOnset = mean_responseOnset, meanDuration = mean_responseDuration,
               sigmaOnset = sdOnset)
simulated_data = simulatePopulation(N=n, mu_O=mean_responseOnset, sigma_O=sdOnset, mu_D_raw=mean_responseDuration) 
#
#
#_________________________________________________________________________________________________________________________________________________________
#
#HYPERPARAMETERS (biased): 
hyperparameters = c(
 		#mean onset, sd mean onset
 		120, 30, 
 		#mean duration, sd mean duration
 		35, 5, 
 		#mean sigma, sd mean sigma
 		5, 2)
#
#_________________________________________________________________________________________________________________________________________________________
#
#PO DATA
responseData.NC = simulated_data$Ts
#
\bold{#STAN RUN PO - type 'intercept-only'}
stanResult.NC = runStanPhenology(
			type="intercept-only",
			responseData=responseData.NC,
		hyperparams_noCovariates=hyperparameters,
			minResponse=0,maxResponse=365,
			runMAP=TRUE,processExtremes=TRUE,N=500,
			partitionDataForPriors=FALSE,
			maxDiv=0,setStringent=TRUE,
			priorLevel=2
			)
#
#SUMMARIZE PO RUN
print(stanResult.NC$sample, max_rows = 15)
##########################################################################################################################################################
}
}
