% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenologyInference.R
\name{SD.CkN}
\alias{SD.CkN}
\title{Title}
\usage{
SD.CkN(
  N,
  mu_O,
  sigma_O,
  mu_D,
  sigma_D = NA,
  minResponse = 0,
  maxResponse = 365,
  type = c("GP", "BB"),
  intFailLow = NA,
  intFailHigh = NA
)
}
\arguments{
\item{N}{}

\item{mu_O}{}

\item{sigma_O}{}

\item{mu_D}{}

\item{sigma_D}{}

\item{minResponse}{}

\item{maxResponse}{}

\item{type}{}

\item{intFailLow}{}

\item{intFailHigh}{}
}
\description{
Title
}
