% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenologyInference.R
\name{SD.Ok1}
\alias{SD.Ok1}
\title{Title}
\usage{
SD.Ok1(
  N,
  mu_O,
  sigma_O,
  minResponse = 0,
  maxResponse = 365,
  intFailLow = NA,
  intFailHigh = NA,
  type = c("GP", "BB")
)
}
\arguments{
\item{N}{}

\item{mu_O}{}

\item{sigma_O}{}

\item{minResponse}{}

\item{maxResponse}{}

\item{intFailLow}{}

\item{intFailHigh}{}

\item{type}{}
}
\description{
Title
}
