% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenologyInference.R
\name{phenologyStandardDeviationFunctions}
\alias{phenologyStandardDeviationFunctions}
\alias{SD.C}
\alias{SD.CkN}
\alias{SD.D}
\alias{SD.O}
\alias{SD.Ok1}
\alias{SD.T}
\title{Functions that calculate the standard deviation (square root of variance) for an associated phenological random variable:}
\usage{
SD.C(
  mu_O = NA,
  sigma_O,
  mu_D = NA,
  sigma_D = NA,
  minResponse = 0,
  maxResponse = 365,
  type = c("GP", "BB")
)

SD.CkN(
  N,
  mu_O,
  sigma_O,
  mu_D,
  sigma_D = NA,
  minResponse = 0,
  maxResponse = 365,
  type = c("GP", "BB"),
  intFailLow = NA,
  intFailHigh = NA
)

SD.D(
  mu_O = NA,
  sigma_O = NA,
  mu_D = NA,
  sigma_D = NA,
  minResponse = 0,
  maxResponse = 365,
  type = c("GP", "BB")
)

SD.O(sigma_O, type = c("GP", "BB"))

SD.Ok1(
  N,
  mu_O,
  sigma_O,
  minResponse = 0,
  maxResponse = 365,
  intFailLow = NA,
  intFailHigh = NA,
  type = c("GP", "BB")
)

SD.T(
  mu_O,
  sigma_O,
  mu_D,
  sigma_D = NA,
  minResponse = 0,
  maxResponse = 365,
  type = c("GP", "BB")
)
}
\arguments{
\item{mu_O}{Mean onset time}

\item{sigma_O}{Standard deviation for the onset time distribution}

\item{mu_D}{Mean phenophase duration length}

\item{sigma_D}{Standard deviation for the phenophase duration distribution}

\item{minResponse}{Minimum value of the response (e.g., day of year); must be set to 0 under current implementation (default = 0)}

\item{maxResponse}{Maximum value of the response (e.g., day of year); typically 365 for Gregorian calendar (default = 365)}

\item{type}{The model type, either BB (beta onset, beta duration) or GP (Gaussian process with a shared standard deviation for onset and cessation and a constant duration) (default = "GP")}

\item{N}{The population size for estimation of extreme events}

\item{intFailLow}{Lowest acceptable integration value (optional; leave at default for most causes)}

\item{intFailHigh}{Highest acceptable integration value (optional; leave at default for most causes)}
}
\value{
The standard deviation of the associated random variable for the input parameter values.
}
\description{
O: phenological onset times
D: phenophase durations
C: phenological cessation times
Ok1: first onset (order k = 1)
CkN: last cessation (order k = N, the population size)
T: observed specimen collection times, e.g., day of year for a yearly time period
PNt: Proportion of the population of size N in the phenophase at time t; only the unnormalized (for efficiency) 'd' and 'r' functions are currently implemented
}
\details{
Most functions share a subset of the same input parameters, but if you use a function that does not require all the parameters, provide only the needed parameters.

The function name is SD.* where * is one of the above random variables.
}
\examples{
#Set the mean onset time
mean_onset = 100
#Set the onset time standard deviation
sigma_onset = 10
#Set the mean duration of the phenophase
mean_duration = 50
#Set the duration standard deviation
sigma_duration = 7
#Set the population size
N=1000
#Calculate the standard deviation of the distribution of last cessation times under the beta onset, beta duration (BB) model
sdCkN = SD.CkN(N=N, mu_O=mean_onset, sigma_O=sigma_onset, mu_D=mean_duration, sigma_D=sigma_duration, type="BB")

}
