% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenologySimulation.R
\name{simulateCorrelatedCovariateAndResponseData}
\alias{simulateCorrelatedCovariateAndResponseData}
\title{Simulate a response variable based on multiple correlated covariates}
\usage{
simulateCorrelatedCovariateAndResponseData(
  n,
  beta,
  cov_names,
  cov_means = NULL,
  response_name = "Y",
  Sigma = NULL,
  R = NULL,
  sds = NULL,
  anchor = 0,
  noise_sd = 1
)
}
\arguments{
\item{n}{Sample size}

\item{beta}{Vector of slope coefficients of the covariates}

\item{cov_names}{Vector of the names of the covariates}

\item{response_name}{The name of the response variable (default: "Y")}

\item{Sigma}{The covariance matrix. Must be of dimension length(beta) X length(beta) (default: identity matrix). Either this or the correlation matrix R with covariate standardiviations (sds) should be provided, and if none, the identity matrix is default. Default: identity matrix}

\item{R}{The correlation matrix. Either with the standard deviations of each covariate (sds) or the Sigma covariance matrix should be given, and if none, the identity matrix is default. Default: identity matrix}

\item{anchor}{Marginal mean value of the response variable}

\item{noise_sd}{Standard deviation of the noise in the response variable}

\item{mu}{Vector of the means of the covariates; must be of same length as beta. (default: all 0 mean)}
}
\value{
A data frame with columns labeled with the variable names and rows representing simulation replicates.
}
\description{
Simulate the values of a response variable based on simulated values of multiple correlated covariates. Assumes a linear model with normally distributed variation.
}
\examples{
\donttest{
#Set the model parameters
covariate_names = c("x1", "x2", "x3") 
slopes = c(1,2,3)
names(slopes) = covariate_names
means = c(10,20,30)
names(means) = covariate_names
response_name = "y"
covariance_matrix = matrix(c( 1.0, 0.5, 0.3, 0.5, 2.0, 0.4, 0.3, 0.4, 1.5), nrow = 3
                           , byrow = TRUE)
colnames(covariance_matrix) = covariate_names
rownames(covariance_matrix) = covariate_names
mean_response = 100
noise = 3
n=50000
#Simulate the data
simulated_data = simulateCorrelatedCovariateAndResponseData(n=n, beta=slopes, cov_names=covariate_names
                                                 , cov_means = means, Sigma=covariance_matrix
                                                 , anchor=mean_response
                                                 , response_name = response_name
                                                 , noise_sd = noise)
#Make a scatter plot of the simulated data
plot(simulated_data$x1, simulated_data$y, main=NULL, xlab="X1", ylab="Y")
##Compare inferences to true values
#coefficients should be 1, 2, 3
summary(lm(simulated_data$y ~ simulated_data$x1 + simulated_data$x2 + simulated_data$x3))
#Means should be 100, 10, 20, 30
colMeans(simulated_data)
}
}
