% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenologySimulation.R
\name{simulateCovariate}
\alias{simulateCovariate}
\title{Title}
\usage{
simulateCovariate(
  n,
  slopeO,
  interceptO,
  sigma,
  slopeD,
  interceptD,
  minCovariate,
  maxCovariate
)
}
\arguments{
\item{n}{}

\item{slopeO}{}

\item{interceptO}{}

\item{sigma}{}

\item{slopeD}{}

\item{interceptD}{}

\item{minCovariate}{}

\item{maxCovariate}{}
}
\description{
Title
}
