% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenologySimulation.R
\name{simulateCovariateSlope}
\alias{simulateCovariateSlope}
\title{Simulate a slope value for a phenological stage response to covariates}
\usage{
simulateCovariateSlope(
  windowBelow = 10,
  windowAbove = 10,
  minCovariate = -10,
  maxCovariate = 10,
  stageMinimumSeparation = 10
)
}
\arguments{
\item{windowBelow}{The mean duration of the stage before the current one. (default: 10)}

\item{windowAbove}{The mean duration of the current stage. (default: 10)}

\item{minCovariate}{The minimum value of the covariate. (default: -10)}

\item{maxCovariate}{The maximum value of the covariate. (default: 10)}

\item{stageMinimumSeparation}{A value describing the minimum amount of separation between the previous and the subsequent stage. (default: 10)}
}
\value{
A slope value giving rise to data that fit, on average, within the input constraints
}
\description{
Simulates a slope value for a phenological stage response to covariates
}
\examples{
\donttest{
previousStageMeanDuration = 50
currentStageMeanDuration = 60
minCovariate = -4
maxCovariate = 27
stageMinimumSeparation = 5
slope = simulateCovariateSlope(previousStageMeanDuration, currentStageMeanDuration, minCovariate, maxCovariate, stageMinimumSeparation=10)
n = 100
noise = 3
x = runif(n, minCovariate,maxCovariate)
y = rnorm(n, previousStageMeanDuration + x * slope, noise) 
plot(x,y)
}
}
