% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenologySimulation.R
\name{simulateFactorCorrelation}
\alias{simulateFactorCorrelation}
\title{Simulate a correlation matrix}
\usage{
simulateFactorCorrelation(
  C = 3,
  K = 2,
  loading_sd = 0.7,
  var_min = 0.3,
  var_max = 1
)
}
\arguments{
\item{C}{The number of covariates (default: 3)}

\item{K}{The number of hidden factors that influence the correlations of the covariates (default: 2)}

\item{loading_sd}{Standard deviation of the strength of covariation. Defines the average difference from the mean covariance, which is 0 (default: 0.7)}

\item{var_min}{The minimum variance of a covariate, which is uniformly sampled between var_min and var_max. Must be positive and less than var_max. (default: 0.3)}

\item{var_max}{The maximum variance of a covariate, which is uniformly sampled between var_min and var_max. Must be positive and more than var_min. (default: 1)}
}
\value{
A list with the correlation matrix, R, the covariance matrix, Sigma, the loading matrix Lambda defined by the normally distributed loading, the diagonal matrix of variances, Psi, covariateSDs, where Sigma is Lambda \%*\% t(Lambda) + Psi
}
\description{
Simulates a correlation matrix, coviariate variances, and provides the covariance matrix as well.
}
\examples{
\donttest{
nCovariates = 5
nLatentFactorsUnderlyingCorrelations = 2
averageDeviationFromMeanCovariance = 0.7
minimumVariance = 0.3
maximumVariance = 1.0
simulateFactorCorrelation(nCovariates, nLatentFactorsUnderlyingCorrelations, averageDeviationFromMeanCovariance, minimumVariance, maximumVariance)
}
}
