% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenologySimulation.R
\name{simulateMultistageData}
\alias{simulateMultistageData}
\title{Simulate covariate data, phenological stage onset times, sample times, and sampled stage data}
\usage{
simulateMultistageData(
  n = 1000,
  nStages = 2,
  stageNames = NULL,
  stage1OnsetMean = NULL,
  stage1OnsetSD = NULL,
  stage1OnsetCovariateSlopes = NULL,
  stageDurationMeans = NULL,
  stageDurationSDs = NULL,
  stageDurationCovariateSlopes = NULL,
  stageMinimumSeparation = 10,
  minStageVariance = 0,
  maxStageVariance = 3,
  meanOnsetSpread = 5,
  minResponse = 0,
  maxResponse = 365,
  nCovariates = 1,
  X = NULL,
  covariateMeans = NULL,
  covariateSDs = NULL,
  covariateNames = NULL,
  R = NULL,
  Sigma = NULL,
  nHiddenFactors = 2,
  hiddenFactorStrength = 0.7,
  minCovariateVariance = 1,
  maxCovariateVariance = 25,
  minCovariateMean = -10,
  maxCovariateMean = 10,
  seed = NULL
)
}
\arguments{
\item{n}{The sample size. (default: 1000)}

\item{nStages}{The number of stages to simulate. Not optional. (default: 2)}

\item{stageNames}{A vector of the names of stages. (default: NULL)}

\item{stage1OnsetMean}{The mean value of the first stage's onset. (default: NULL)}

\item{stage1OnsetSD}{The standard deviation of the first stage's onset. (default: NULL)}

\item{stage1OnsetCovariateSlopes}{A vector with the first stage's onset model coefficients. (default: NULL)}

\item{stageDurationMeans}{A vector with all but the last stage's mean durations. (default: NULL)}

\item{stageDurationSDs}{A vector with all but the last stage's duration standard deviations. (default: NULL)}

\item{stageDurationCovariateSlopes}{A matrix with the slopes for each coefficient of the duration model for each stage. Columns are the covariates, and rows are the stages. (default: NULL)}

\item{stageMinimumSeparation}{The minimum separation, on average, between stage onsets. (default: 10)}

\item{minStageVariance}{The minimum allowable variance in durations for a stage. (default: 0)}

\item{maxStageVariance}{The maximum allowable variance in durations for a stage. (default: 3.0)}

\item{meanOnsetSpread}{A value or a vector of values (one for each stage), indicating the relative duration, on average, of that stage. Positive numbers only. Smaller numbers result in greater variation of durations. (default: 5.0)}

\item{minResponse}{The minimum response time. (default: 0)}

\item{maxResponse}{The maximum response time. (default: 365)}

\item{nCovariates}{The number of covariates. (default: 1)}

\item{X}{An n X nCovariates matrix with the covariate data. (default: NULL)}

\item{covariateMeans}{A vector of covariate means. (default: NULL)}

\item{covariateSDs}{A vector of covariate standard deviations. (default: NULL)}

\item{covariateNames}{A vector of covariate names. (default: NULL)}

\item{R}{A correlation matrix for the covariates. Used if no Sigma covariance matrix is provided. (default: NULL)}

\item{Sigma}{A covariance matrix for the covariates. (default: NULL)}

\item{nHiddenFactors}{The number of hidden factors that link to the covariates, if their correlation structure is to be simulated. (default: 2)}

\item{hiddenFactorStrength}{The average deviation from the mean covariance of 0. (default: 0.7)}

\item{minCovariateVariance}{The minimum allowable variance of a covariate. (default: 1.0)}

\item{maxCovariateVariance}{The maximum allowable variance of a covariate. (default: 25.0)}

\item{minCovariateMean}{The minimum mean value of a covariate (default: -10.0)}

\item{maxCovariateMean}{The maximum mean value of a covariate (default: 10.0)}

\item{seed}{The random number generator seed. (default: NULL)}
}
\value{
A data frame with the covariate data, the stage onset times, the sampled times, and the stage at the sampled times
}
\description{
Simulate covariate data, phenological stage onset times, sample times, and sampled stage data.
}
\examples{
\donttest{
plotSimulation = function(data, target, colors,nCovariates) {
cols = colors
for(i in 1:numberStages) {
 if(i == 1) {
	plot(data$outputData[,target],data$outputData[,nCovariates+1],col=cols[1],ylim=c(0,365),pch=16)
}
else {
	points(data$outputData[,target],data$outputData[,nCovariates+i],col=cols[i],pch=16)
}
fit = lm(data$outputData[,nCovariates+i] ~ data$outputData[,1])
abline(fit,col=cols[i])
}
segments(x0=data$outputData[,target],x1=data$outputData[,target],y0=0,y1=365,col="gray25")
points(data$outputData[,target], data$outputData$sampledTime, col=cols[data$outputData$sampledStage], pch=4, cex=3)
}

numberStages = 5
numberCovariates = 3
sampleSize = 50
simulatedData = simulateMultistageData(n=sampleSize, nStages=numberStages, nCovariates=numberCovariates)
targetCovariate = 1
cols = c("red","blue","orange","black","pink")
plotSimulation(simulatedData, targetCovariate, cols,numberCovariates)
}
}
