% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenologySimulation.R
\name{simulatePopulation}
\alias{simulatePopulation}
\title{Title}
\usage{
simulatePopulation(
  N,
  mu_O,
  sigma_O,
  mu_D_raw,
  sigma_D = NA,
  minResponse = 0,
  maxResponse = 365,
  mins = 1.5,
  maxs = 3000,
  type = "GP"
)
}
\arguments{
\item{N}{}

\item{mu_O}{}

\item{sigma_O}{}

\item{mu_D_raw}{}

\item{sigma_D}{}

\item{minResponse}{}

\item{maxResponse}{}

\item{mins}{}

\item{maxs}{}

\item{type}{}
}
\description{
Title
}
