% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenologySimulation.R
\name{simulateResponseData}
\alias{simulateResponseData}
\title{Simulate a response variable based on a linear model of covariates}
\usage{
simulateResponseData(X, beta, response_name = "Y", anchor = 0, noise_sd = 0)
}
\arguments{
\item{X}{A data frame with labeled columns indicating covariate names and rows with covariate data}

\item{beta}{Vector of slope coefficients of the covariates}

\item{response_name}{The name of the response variable (default: "Y")}

\item{anchor}{Marginal mean value of the response variable}

\item{noise_sd}{Standard deviation of the noise in the response variable}

\item{Sigma}{The covariance matrix. Must be of dimension length(beta) X length(beta) (default: identity matrix)}
}
\value{
A data frame with columns labeled with the variable names and rows representing simulation replicates.
}
\description{
Simulate the values of a response variable based on values of multiple covariates. Assumes a linear model with normally distributed variation.
}
\examples{
\donttest{
##First, simulate covariate data
#Set the model parameters for covariates
covariate_names = c("x1", "x2", "x3") 
means = c(10,20,30)
names(means) = covariate_names
correlation_matrix = matrix(c( 1.0, 0.5, 0.3, 
			 0.5, 1.0, 0.4, 
			 0.3, 0.4, 1), nrow = 3
                           , byrow = TRUE)
rownames(correlation_matrix) = covariate_names
colnames(correlation_matrix) = covariate_names
sds = c(1,2,4)
names(sds) = covariate_names
n=1000
X = simulateCorrelatedCovariateData(n=n, cov_names=covariate_names, means=means, R=correlation_matrix, sds=sds)

##Next, simulate response data based on simulated covariate data
#Set the model parameters
slopes = c(1,2,3)
names(slopes) = covariate_names
response_name = "y"
mean_response = 100
noise = 3
#Simulate the data
simulated_data = simulateResponseData(X=X, beta=slopes, 
                                                 response_name = response_name,
                                                 anchor=mean_response,
                                                 noise_sd = noise)
#Make a scatter plot of the simulated data
plot(simulated_data$x1, simulated_data$y, main=NULL, xlab="X1", ylab="Y")
#Present some stats to check match with true parameters
#Coefficients should match slopes for the covariates
summary(lm(simulated_data$y ~ simulated_data$x1 + simulated_data$x2 + simulated_data$x3))
#Correlation matrix should be close to the true correlation matrix
cor(X)
#Means should resemble the provided ones for the response and covariates
colMeans(simulated_data)
}
}
