% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{summarizePhenologyResults}
\alias{summarizePhenologyResults}
\title{Title}
\usage{
summarizePhenologyResults(
  stanRunResult,
  taxonName,
  measures = c("mean", "median", "sd", "mad"),
  quantiles = c(2.5, 97.5),
  convergence = c("rhat", "ess_bulk", "ess_tail"),
  standardLinearModel = NA,
  processExtremes = TRUE,
  N = 500
)
}
\arguments{
\item{stanRunResult}{}

\item{taxonName}{}

\item{measures}{}

\item{quantiles}{}

\item{convergence}{}

\item{standardLinearModel}{}

\item{processExtremes}{}

\item{N}{}
}
\description{
Title
}
