% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{summaryStanDiagnostics}
\alias{summaryStanDiagnostics}
\title{Summary diagnostics of a Stan run}
\usage{
summaryStanDiagnostics(
  stanRunResult,
  NTotal,
  taxonName,
  diagnostics = c("num_divergent", "num_max_treedepth", "ebfmi")
)
}
\arguments{
\item{stanRunResult}{The output from the runStanPhenology function}

\item{NTotal}{The sample size (i.e., number of specimens used in analysis)}

\item{taxonName}{The taxon name}

\item{diagnostics}{A vector of named diagnostic quantities (default: c("num_divergent", "num_max_treedepth", "ebfmi"))}
}
\value{
A list with the taxon name, sample size total, sample size used by Stan, and a data frame with the Stan diagnostics summary table
}
\description{
Provides a data frame with diagnostic information for a Stan run.
}
