% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenologyDistributions.R
\name{phenologyDistributionFunctions}
\alias{phenologyDistributionFunctions}
\alias{dO}
\alias{pO}
\alias{qO}
\alias{rO}
\alias{dD}
\alias{pD}
\alias{rD}
\alias{qD}
\alias{dOk1}
\alias{qOk1}
\alias{pOk1}
\alias{rOk1}
\alias{dC}
\alias{pC}
\alias{rC}
\alias{qC}
\alias{dT}
\alias{pT}
\alias{qT}
\alias{rT}
\alias{rCkN}
\alias{dCkN}
\alias{pCkN}
\alias{qCkN}
\alias{dR}
\alias{rR}
\alias{dPNt}
\alias{rPNt}
\title{Functions that model phenological distributions.}
\usage{
dO(x, mu_O, sigma_O, minResponse = 0, maxResponse = 365, type = c("GP", "BB"))

pO(q, mu_O, sigma_O, minResponse = 0, maxResponse = 365, type = c("GP", "BB"))

qO(p, mu_O, sigma_O, minResponse = 0, maxResponse = 365, type = c("GP", "BB"))

rO(n, mu_O, sigma_O, minResponse = 0, maxResponse = 365, type = c("GP", "BB"))

dD(
  x,
  mu_O = NA,
  sigma_O = NA,
  mu_D,
  sigma_D = NA,
  minResponse = 0,
  maxResponse = 365,
  type = c("GP", "BB")
)

pD(
  q,
  mu_O = NA,
  sigma_O = NA,
  mu_D,
  sigma_D = NA,
  minResponse = 0,
  maxResponse = 365,
  type = c("GP", "BB")
)

rD(
  n,
  mu_O = NA,
  sigma_O = NA,
  mu_D,
  sigma_D = NA,
  minResponse = 0,
  maxResponse = 365,
  type = c("GP", "BB")
)

qD(
  p,
  mu_O = NA,
  sigma_O = NA,
  mu_D,
  sigma_D = NA,
  minResponse = 0,
  maxResponse = 365,
  type = c("GP", "BB")
)

dOk1(
  x,
  N,
  mu_O,
  sigma_O,
  minResponse = 0,
  maxResponse = 365,
  type = c("GP", "BB")
)

qOk1(
  p,
  N,
  mu_O,
  sigma_O,
  minResponse = 0,
  maxResponse = 365,
  type = c("GP", "BB")
)

pOk1(
  q,
  N,
  mu_O,
  sigma_O,
  minResponse = 0,
  maxResponse = 365,
  type = c("GP", "BB")
)

rOk1(
  n,
  N,
  mu_O,
  sigma_O,
  minResponse = 0,
  maxResponse = 365,
  type = c("GP", "BB")
)

dC(
  x,
  mu_O,
  sigma_O,
  mu_D,
  sigma_D,
  minResponse = 0,
  maxResponse = 365,
  type = c("GP", "BB")
)

pC(
  q,
  mu_O,
  sigma_O,
  mu_D,
  sigma_D,
  minResponse = 0,
  maxResponse = 365,
  type = c("GP", "BB")
)

rC(
  n,
  mu_O,
  sigma_O,
  mu_D,
  sigma_D,
  minResponse = 0,
  maxResponse = 365,
  type = c("GP", "BB")
)

qC(
  p,
  mu_O,
  sigma_O,
  mu_D,
  sigma_D,
  minResponse = 0,
  maxResponse = 365,
  type = c("GP", "BB")
)

dT(
  x,
  mu_O,
  sigma_O,
  mu_D,
  sigma_D,
  minResponse = 0,
  maxResponse = 365,
  type = c("GP", "BB")
)

pT(
  q,
  mu_O,
  sigma_O,
  mu_D,
  sigma_D,
  minResponse = 0,
  maxResponse = 365,
  type = c("GP", "BB")
)

qT(
  p,
  mu_O,
  sigma_O,
  mu_D,
  sigma_D,
  minResponse = 0,
  maxResponse = 365,
  type = c("GP", "BB")
)

rT(
  n,
  mu_O,
  sigma_O,
  mu_D,
  sigma_D,
  minResponse = 0,
  maxResponse = 365,
  type = c("GP", "BB")
)

rCkN(
  n,
  N,
  mu_O,
  sigma_O,
  mu_D,
  sigma_D,
  minResponse = 0,
  maxResponse = 365,
  type = c("GP", "BB")
)

dCkN(
  x,
  N,
  mu_O,
  sigma_O,
  mu_D,
  sigma_D,
  minResponse = 0,
  maxResponse = 365,
  type = c("GP", "BB")
)

pCkN(
  q,
  N,
  mu_O,
  sigma_O,
  mu_D,
  sigma_D,
  minResponse = 0,
  maxResponse = 365,
  type = c("GP", "BB")
)

qCkN(
  p,
  N,
  mu_O,
  sigma_O,
  mu_D,
  sigma_D,
  minResponse = 0,
  maxResponse = 365,
  type = c("GP", "BB")
)

dR(
  x,
  N,
  mu_O,
  sigma_O,
  mu_D,
  sigma_D,
  minResponse = 0,
  maxResponse = 365,
  type = c("GP", "BB")
)

rR(
  n,
  N,
  mu_O,
  sigma_O,
  mu_D,
  sigma_D,
  minResponse = 0,
  maxResponse = 365,
  type = c("GP", "BB")
)

dPNt(
  x,
  t,
  mu_O,
  sigma_O,
  mu_D,
  sigma_D,
  minResponse = 0,
  maxResponse = 365,
  type = c("GP", "BB")
)

rPNt(
  n,
  t,
  mu_O,
  sigma_O,
  mu_D,
  sigma_D,
  minResponse = 0,
  maxResponse = 365,
  type = c("GP", "BB")
)
}
\arguments{
\item{x}{A vector of the response data (e.g. day of year values)}

\item{mu_O}{Mean onset time}

\item{sigma_O}{Standard deviation for the onset time distribution}

\item{minResponse}{Minimum value of the response (e.g., day of year); must be set to 0 under current implementation (default = 0)}

\item{maxResponse}{Maximum value of the response (e.g., day of year); typically 365 for Gregorian calendar (default = 365)}

\item{type}{The model type, either BB (beta onset, beta duration) or GP (Gaussian process with a shared standard deviation for onset and cessation and a constant duration) (default = "GP")}

\item{q}{The quantile}

\item{p}{The lower tail area (probability)}

\item{n}{The sample size}

\item{mu_D}{Mean phenophase duration length}

\item{sigma_D}{Standard deviation for the phenophase duration distribution}

\item{N}{The population size}

\item{t}{The time at which the proportion of the population in the phenophase is to be evaluated}
}
\value{
A vector, following standard R conventions for the 'd', 'p', 'r', and 'q' functions
}
\description{
Methods to model and simulate phenological distributions. The R convention of 'd' (density or mass function), 'r' (sampling), 'q' (quantile), and 'p' (cumulative distribution function) prefixes the function name.

Eight phenological properties are modeled by random variables. These include the following:

O: phenological onset times

D: phenophase durations

C: phenological cessation times

Ok1: first onset (order k = 1)

CkN: last cessation (order k = N, the population size)

T: observed specimen collection times, e.g., day of year for a yearly time period

PNt: Proportion of the population of size N in the phenophase at time t; only the unnormalized 'd' and random sampling 'r' functions are currently implemented

R: The total time range that at least one individual in a population is in the phenophase; only the 'd' and 'r' functions are currently implemented

The function name is constructed by combining one of 'd', 'r', 'q', and 'p' as a prefix with one of the above variable labels. For example, rT refers to the function that generates random samples from the distribution of observed collection times, T.

Most functions share a subset of input parameters. Provide only the needed parameters for the function.
}
\examples{
\donttest{
#Set the sample size
n=100000
#Set the mean onset time
mean_onset = 100
#Set the onset time standard deviation, sigma
sigma_onset = 10
#Set the duration of the phenophase
duration = 50
#Sample the observed collection times
observed_t = rT(n=n, mu_O = mean_onset, sigma_O=sigma_onset, mu_D=duration)
#Make a histogram of the observed collection times
hist(observed_t, probability=TRUE, xlab="Simulated observed collection times")
#Overlay the theoretical curve on the histogram
curve(dT(x,mu_O=mean_onset,sigma_O=sigma_onset, mu_D=duration),col="red",add=TRUE)

#Or, for a more "extreme" distribution that illustrates the flexibility of the model's shape:
#Set the mean onset time
mean_onset = 180
#Set the onset time standard deviation, sigma
sigma_onset = 100
#Set the duration of the phenophase
mean_duration = 50
sigma_duration = 10
#Sample the observed collection times
observed_t = rT(n=n, mu_O = mean_onset, sigma_O=sigma_onset, mu_D=mean_duration
                , sigma_D=sigma_duration, type="BB")
#Make a histogram of the observed collection times
hist(observed_t, probability=TRUE
     , xlab="Simulated observed collection times (extreme BB model)")
#Overlay the theoretical curve on the histogram
curve(dT(x,mu_O=mean_onset,sigma_O=sigma_onset, mu_D=mean_duration
      , sigma_D=sigma_duration, type="BB"),col="red",add=TRUE,from=0, to=365)
}
}
