% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{getHyperparametersViaQuantileRegression}
\alias{getHyperparametersViaQuantileRegression}
\title{Title}
\usage{
getHyperparametersViaQuantileRegression(
  responseDataForPrior,
  onsetCovariateDataForPrior,
  durationCovariateDataForPrior,
  lowerQuantile = 0.1,
  upperQuantile = 0.9,
  confidence = 2,
  scale = 1
)
}
\arguments{
\item{responseDataForPrior}{}

\item{onsetCovariateDataForPrior}{}

\item{durationCovariateDataForPrior}{}

\item{lowerQuantile}{}

\item{upperQuantile}{}

\item{confidence}{}

\item{scale}{}
}
\description{
Title
}
