% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenologyInference.R
\name{getMAP}
\alias{getMAP}
\title{Title}
\usage{
getMAP(
  responseData,
  minResponse = 0,
  maxResponse = 365,
  minS = 1,
  maxS = 3000,
  init_params = c(180, 20, 60, 7),
  hyperparameters = c(100, 7, 60, 6, 24, 12, 24, 12),
  type = "BB"
)
}
\arguments{
\item{responseData}{}

\item{minResponse}{}

\item{maxResponse}{}

\item{minS}{}

\item{maxS}{}

\item{init_params}{}

\item{hyperparameters}{}

\item{type}{}
}
\description{
Title
}
