% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenologyInference.R
\name{getPeakPhenophase}
\alias{getPeakPhenophase}
\title{Title}
\usage{
getPeakPhenophase(
  mu_O,
  sigma_O = NA,
  mu_D,
  sigma_D = NA,
  minResponse = 0,
  maxResponse = 365,
  type = c("GP", "BB")
)
}
\arguments{
\item{mu_O}{}

\item{sigma_O}{}

\item{mu_D}{}

\item{sigma_D}{}

\item{minResponse}{}

\item{maxResponse}{}

\item{type}{}
}
\description{
Title
}
