% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenologyVisualization.R
\name{makePosteriorPredictivePlot}
\alias{makePosteriorPredictivePlot}
\title{Title}
\usage{
makePosteriorPredictivePlot(
  stanResult,
  responseData,
  responseVariableName = "DOY",
  targetCovariateName,
  onsetCovariateData,
  durationCovariateData,
  n_samples = 100,
  n_draws = 100,
  n_hist = 10000,
  n_bin = 100,
  resolution = 50,
  slice = 0.25,
  N = 1000,
  minResponse = 0,
  maxResponse = 365,
  adjustYLim = T,
  smooth = c("GAM", "LOESS", "SPLINE", "NONE"),
  keepSigma = T,
  makeHistograms = FALSE
)
}
\arguments{
\item{stanResult}{}

\item{responseData}{}

\item{responseVariableName}{}

\item{targetCovariateName}{}

\item{onsetCovariateData}{}

\item{durationCovariateData}{}

\item{n_samples}{}

\item{n_draws}{}

\item{n_hist}{}

\item{n_bin}{}

\item{resolution}{}

\item{slice}{}

\item{N}{}

\item{minResponse}{}

\item{maxResponse}{}

\item{adjustYLim}{}

\item{smooth}{}

\item{keepSigma}{}

\item{makeHistograms}{}
}
\description{
Title
}
