% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{partitionResponseCovariateData}
\alias{partitionResponseCovariateData}
\title{Title}
\usage{
partitionResponseCovariateData(
  responseData,
  onsetCovariateData,
  durationCovariateData,
  prop = 0.3
)
}
\arguments{
\item{responseData}{}

\item{onsetCovariateData}{}

\item{durationCovariateData}{}

\item{prop}{}
}
\description{
Title
}
