% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{partitionResponseData}
\alias{partitionResponseData}
\title{Title}
\usage{
partitionResponseData(responseData, prop = 0.3)
}
\arguments{
\item{responseData}{}

\item{prop}{}
}
\description{
Title
}
