% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenologyVisualization.R
\name{plotMultistageSimulation}
\alias{plotMultistageSimulation}
\title{Make a plot of simulated multistage data}
\usage{
plotMultistageSimulation(
  simulatedData = NULL,
  targetCovariateIndex = 1,
  stageColors = NULL
)
}
\arguments{
\item{simulatedData}{The output object from the function simulateMultistageData. (default: NULL)}

\item{targetCovariateIndex}{The index number of the covariate to be plotted along the x-axis. (default: 1)}

\item{stageColors}{A vector of colors, one per stage. (default: NULL)}
}
\value{
A plot illustrating the simulated multistage data for one target covariate along the x-axis.
}
\description{
Make a plot of simulated multistage data
}
\examples{
\donttest{
##Set basic simulation parameters
numberStages = 3
numberCovariates = 1
sampleSize = 50

##Simulate data
simulatedData = simulateMultistageData(n=sampleSize, nStages=numberStages, nCovariates=numberCovariates)

##Plot simulated data with the first covariate (the only covariate in this case) along the x-axis.
#    set colors for the stages
stageColors = viridisLite::viridis(numberStages)
# stageColors = rainbow(numberStages)
#    create the plot
plotMultistageSimulation(simulatedData=simulatedData, targetCovariateIndex=1, stageColors=stageColors)
}
}
