% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{preparePhenologyData}
\alias{preparePhenologyData}
\title{Prepare data for analysis}
\usage{
preparePhenologyData(
  dataFile,
  responseVariableName,
  onsetCovariateNames,
  durationCovariateNames,
  removeDuplicateRows = TRUE,
  removeOutliers = FALSE,
  removeIncomplete = TRUE,
  dataSummaryDirectory = NA,
  taxonName = NA,
  origN = NA
)
}
\arguments{
\item{dataFile}{A path to a file with the data. See 'Details' for the correct configuration of this file.}

\item{responseVariableName}{The name of the response variable. Must match the name of a column in 'dataFile'.}

\item{onsetCovariateNames}{A vector with the names of the onset covariates. Must match name(s) of column(s) in 'dataFile'.}

\item{durationCovariateNames}{A vector with the names of the duration covariates. Must match name(s) of column(s) in 'dataFile'.}

\item{removeDuplicateRows}{If duplicated rows should be removed (default: TRUE).}

\item{removeOutliers}{If outliers should be removed (default: TRUE).}

\item{removeIncomplete}{If rows with missing data should be removed (default: TRUE).}

\item{dataSummaryDirectory}{Optionally provide the name of a directory where a summary of the data will be saved in a tab-delimited file named dataSummary.\if{html}{\out{<taxonName>}}.txt}

\item{taxonName}{Optionally, the name of the taxon to be stored in the data summary file.}

\item{origN}{Optionally, the initial number of specimens before any filtering of specimens took place (Use only if creating optional data summary file).}
}
\value{
Returns a list object with elements necessary to use the "runStanPhenology" function. Elements include the original data (originalData) and the cleaned response data (responseData), onset covariate data (onsetCovariateData), and duration covariate data (durationCovariateData). The output data remain unscaled. The runStanPhenology function will scale data appropraitely.
}
\description{
This function prepares the phenological data for Bayesian estimation analysis using "runStanPhenology". The main inputs are the path to the data file and variable names. Note that all variable names should match the names of the columns in file.
}
\details{
The file whose name is provided as input must be a text-formatted, tab-delimited spreadsheet. Each row corresponds to a single specimen, and the columns provide the specimen data. Any number of columns is fine, but the input names of the response variable, the onset covariates, and the duration covariates must match the corresponding header names in the spreadsheet.
}
\examples{
\donttest{
##get the file name with data for the blood root plant
file  =  system.file("data", "Sanguinaria_canadensis.Full.txt", package = "phenoCollectR")
##define the covariate names - remove up to all but 1
vars = c("Latitude", "Year", "Elevation", "AnnualMonthlyAverageTemp", "SpringMonthlyAverageTemp", "FirstQuarterMonthlyAverageTemp")
##get the phenology data
data  =  preparePhenologyData(dataFile=file, responseVariableName="DOY", onsetCovariateNames=vars, durationCovariateNames=vars, taxonName="Sanguinaria_canadensis", removeOutliers=TRUE)
##run the Stan sampler
stanResult  =  runStanPhenology(type="full", responseData = data$responseData, onsetCovariateData = data$onsetCovariateData, durationCovariateData = data$durationCovariateData, partitionDataForPriors = TRUE)
##summarize the Stan run
stanSummary  =  summarizePhenologyResults(stanRunResult = stanResult, taxonName = "Sanguinaria_canadensis",standardLinearModel = TRUE)
##make posterior predictive graph
pp  =  makePosteriorPredictivePlot(stanResult = stanResult, responseData = data$responseData, targetCovariateName = "SpringMonthlyAverageTemp", onsetCovariateData = data$onsetCovariateData, durationCovariateData = data$durationCovariateData)
##display the posterior predictive graph		
print(pp)		
}
}
