% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenologyInference.R
\name{runStandardLinearModel}
\alias{runStandardLinearModel}
\title{Title}
\usage{
runStandardLinearModel(responseData, onsetCovariateData, durationCovariateData)
}
\arguments{
\item{responseData}{}

\item{onsetCovariateData}{}

\item{durationCovariateData}{}
}
\description{
Title
}
