% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenologyInference.R
\name{runStanPhenology}
\alias{runStanPhenology}
\title{Title}
\usage{
runStanPhenology(
  type = c("intercept-only", "full"),
  responseData,
  hyperparams_noCovariates = NA,
  onsetCovariateData = NA,
  durationCovariateData = NA,
  onsetHyperBeta = NA,
  onsetHyperAnchor = NA,
  durationHyperBeta = NA,
  durationHyperAnchor = NA,
  cessationHyperAnchor = NA,
  sigmaHyper = NA,
  minResponse = 0,
  maxResponse = 365,
  maxDiv = 0,
  setStringent = TRUE,
  runMAP = FALSE,
  processExtremes = TRUE,
  N = 500,
  partitionDataForPriors = TRUE,
  maximizeSampleSize = FALSE,
  threshApprox = NA,
  byPassChecks = FALSE,
  ...
)
}
\arguments{
\item{type}{}

\item{responseData}{}

\item{hyperparams_noCovariates}{}

\item{onsetCovariateData}{}

\item{durationCovariateData}{}

\item{onsetHyperBeta}{}

\item{onsetHyperAnchor}{}

\item{durationHyperBeta}{}

\item{durationHyperAnchor}{}

\item{cessationHyperAnchor}{}

\item{sigmaHyper}{}

\item{minResponse}{}

\item{maxResponse}{}

\item{maxDiv}{}

\item{setStringent}{}

\item{runMAP}{}

\item{processExtremes}{}

\item{N}{}

\item{partitionDataForPriors}{}

\item{maximizeSampleSize}{}

\item{threshApprox}{}

\item{byPassChecks}{}

\item{...}{}
}
\description{
Title
}
