% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenologySimulation.R
\name{simulateCorrelatedCovariateData}
\alias{simulateCorrelatedCovariateData}
\title{Title}
\usage{
simulateCorrelatedCovariateData(
  n,
  beta,
  cov_names,
  response_name = "Y",
  Sigma = NULL,
  anchor = 0,
  noise_sd = 1
)
}
\arguments{
\item{n}{}

\item{beta}{}

\item{cov_names}{}

\item{response_name}{}

\item{Sigma}{}

\item{anchor}{}

\item{noise_sd}{}
}
\description{
Title
}
