% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenologySimulation.R
\name{simulateMultistageOverlapData}
\alias{simulateMultistageOverlapData}
\title{Simulate covariate data and phenological unit stages nested within individuals}
\usage{
simulateMultistageOverlapData(
  n = 500,
  meanUnits = 20,
  nPre = 1,
  nVisible = 2,
  nPost = 1,
  nCovariates = 2,
  ...
)
}
\arguments{
\item{n}{The sample size. (default: 500)}

\item{meanUnits}{The mean number of phenological units that an individual develops during a phenological time period (e.g., an individual plant produces multiple flower units during a phenological cycle). (default: 20)}

\item{nPre}{The number of stages before overlapping phenological units (e.g., floral units) begin development}

\item{nVisible}{The number of stages for which overlapping phenological units (e.g., floral units) are present (e.g., bud, flower, immature fruit, mature fruit)}

\item{nPost}{The number of stages after all overlapping phenological units (e.g., floral units) have abscised (after fruits have fallen off of plant)}

\item{nCovariates}{The number of covariates to simulate. (default: 2)}

\item{...}{Additional optional parameters passed to simulateMultistageData (called from within this simulateMultistageOverlapData function).}
}
\value{
A list with the following:

nPre: the number of stages before (pre, such as dormant stage, pre-reproductive vegetative stage) development of overlapping phenological units (such as flower units)
nVisible: the number of stages when overlapping phenological units are visible (such as bud, flower, immature fruit, mature fruit for floral units)
nPost: the number of stages after (post) all phenological units have absciced from the individual.
Data frame named 'simulatedData' with the output from the internally-called simuateMultistageData function
A list of 'simulatedIndividuals' with:
The time when an individual was observed
A vector, 'phenologicalUnits', of counts of phenological units for each stage that includes latent as well as observed units
A vector, 'stageCounts', of the counts of observed phenological units within each of the nStages stages. If the individual is in a pre- or post- stage, there will be a 1 in the index position of that stage and zeroes in the other stages. Otherwise, there will be a count of the number of observed phenological units in the stage and 0's in all the pre- and post- stages.
A vector, 'pi', of the probabilities of being in each stage at the sampled time.
nUnits: the total number of units (latent and observed) for the individual. It is possible that not all have developed yet and it is possible that some have already abscised.
}
\description{
Simulate covariate data, phenological unit stage counts, onset times, and sampled times. An example would be an individual plant with multiple flowers as the phenological unit, with flower units in possibly different phenological stages (bud, flower, immature fruit, mature fruit). The function also optionally models transitions from a pre-(phenological unit) stage, to stages with multiple phenological units, to post-(phenological unit) stages after those units have senesced.
}
\examples{
\donttest{
##Set basic simulation parameters 
nPre = 1           
nVisible = 3
nPost = 1
numberCovariates = 3
popSize = 500	      #Number of individuals to simulate in the simulated population
meanUnits = 30             #Set the mean number of phenological units (e.g., flowers) that an individual develops during a phenological cycle.

##Simulate data
sim = simulateMultistageOverlapData(n=popSize, nPre = nPre, nVisible = nVisible, nPost = nPost, meanUnits=meanUnits, nStages=numberStages, nCovariates=numberCovariates)

##Plot simulated data with the first covariate along the x-axis and the predominant stage color coded. Note, this plot does not represent all the stages of all the phenological units, but rather a single 'x' for each indivdiual color-coded by the predominant stage.
#	set colors for the stages
#stageColors = rainbow(numberStages) 
#stageColors = viridisLite::viridis(numberStages)
stageColors = c("#332288", "#88CCEE", "#44AA99", "#117733", "#999933", "#DDCC77", "#CC6677", "#882255", "#AA4499") #Paul Tol
#	create the plot
plotMultistageSimulation(simulatedData=sim$simulatedData, targetCovariateIndex=1, stageColors=stageColors)
}
}
