% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenologySimulation.R
\name{simulatePopulationLatentIntervalStates}
\alias{simulatePopulationLatentIntervalStates}
\title{Simulate a population of individuals with simulated onset, duration, cessation, and observed times based on multiple correlated covariates for onset and for duration}
\usage{
simulatePopulationLatentIntervalStates(
  n,
  betaOnset = 0,
  betaDuration = 0,
  covariateNamesOnset = "noOnsetCovariates",
  covariateNamesDuration = "noDurationCovariates",
  muCovariateOnset = 0,
  muCovariateDuration = 0,
  CovarianceOnset = 0,
  CovarianceDuration = 0,
  meanOnset = 0,
  meanDuration = 0,
  sigmaOnset = 1,
  minResponse = 0,
  maxResponse = 365
)
}
\arguments{
\item{n}{Sample size}

\item{betaOnset}{Vector of slope coefficients of the covariates for onset.}

\item{betaDuration}{Vector of slope coefficients of the covariates for duration.}

\item{covariateNamesOnset}{Vector of the names of the covariates for onset; must be same length as betaOnset (default: "noOnsetCovariates").}

\item{covariateNamesDuration}{Vector of the names of the covariates for duration; must be same length as betaDuration (default: "noDurationCovariates").}

\item{muCovariateOnset}{Vector of the means of the covariates for onset; must be of same length as betaOnset. Helps to define the intercept. (default: 0)}

\item{muCovariateDuration}{Vector of the means of the covariates for duration; must be of same length as betaDuration. Helps to define the intercept. (default: 0)}

\item{CovarianceOnset}{The covariance matrix for onset; must be of dimension length(betaOnset) X length(betaOnset) (default: 0)}

\item{CovarianceDuration}{The covariance matrix for duration; must be of dimension length(betaDuration) X length(betaDuration) (default: 0)}

\item{meanOnset}{Marginal mean value of the onset times.}

\item{meanDuration}{Marginal mean value of the duration times.}

\item{sigmaOnset}{Standard deviation of the onset times. Duration is assumed fixed in this model.}

\item{minResponse}{The minimum time of observations. default: 0, which is the only fully supported value.}

\item{maxResponse}{The maximum time of observations. default: 365, representing the number of days in a year.}
}
\value{
Dataframe with columns labeled with the variable names, including all covariates, the simulated onset times, durations, cessations, sampled times, state (before, during, after phenophase) and rows representing simulated individuals
}
\description{
Simulate a population of individuals with simulated onset, duration, cessation, and observed times based on multiple correlated covariates for onset and for duration. Assumes a linear model with normally distributed variation. Defaults produced simulated values with no covariates.
}
\examples{
\donttest{
#Set the model parameters
slopesOnset = c(1,2,3)
meansOnset = c(10,20,30)
covariate_namesOnset = c("o1", "o2", "o3") 
response_nameOnset = "onset"
covariance_matrixOnset = matrix(c( 1.0, 0.5, 0.3, 0.5, 2.0, 0.4, 0.3, 0.4, 1.5), nrow = 3
                           , byrow = TRUE)
mean_responseOnset = 150
noiseOnset = 3
slopesDuration = c(1,3)
meansDuration = c(50,40)
covariate_namesDuration = c("d1", "d2") 
response_nameDuration = "duration"
covariance_matrixDuration = matrix(c( 0.5, 0.5, 0.3, 0.5), nrow = 2
                           , byrow = TRUE)
mean_responseDuration = 30
n=1000
#Simulate the data
simulated_data = simulatePopulationLatentIntervalStates(n=n,
							betaOnset=slopesOnset, betaDuration=slopesDuration,
						covariateNamesOnset=covariance_namesOnset, covariateNamesDuration=covariance_namesDuration,
						muOnset = meansOnset, muDuration = meansDuration,
						SigmaOnset = covariance_matrixOnset, SigmaDuration = covariance_matrixDuration,
						meanOnset = mean_responseOnset, meanDuration = mean_responseDuration,
						sigmaOnset = noiseOnset)
#Make a scatter plot of the simulated data
plot(simulated_data$o1, simulated_data$onset, main=NULL, xlab="O1", ylab="Onset")
}
}
