% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{summaryStanDiagnostics}
\alias{summaryStanDiagnostics}
\title{Title}
\usage{
summaryStanDiagnostics(
  stanRunResult,
  NTotal,
  taxonName,
  diagnostics = c("num_divergent", "num_max_treedepth", "ebfmi")
)
}
\arguments{
\item{stanRunResult}{}

\item{NTotal}{}

\item{taxonName}{}

\item{diagnostics}{}
}
\description{
Title
}
